Form clusters of datapoints using Hartigan's Leader Algorithm.

Form clusters of datapoints using Hartigan's Leader Algorithm. This is a modification of getHDmembers.

get_leader_cluster(data, maxrows = 1000)

Arguments

data
A vector, matrix, or data frame consisting of numeric and/or categorical variables.

maxrows
If the number of observations is greater than maxrows, outliers reduces the number used in k-nearest-neighbor computations to a set of exemplars. The default value is 1000.

Value

A list in which each component is a vector of observation indexes. The first index in each list is the index of the exemplar defining that list, and any remaining indexes are the associated members, within radius of the exemplar.

References

Hartigan, John A. "Clustering algorithms." (1975). Kantardzic, Mehmed. Data mining: concepts, models, methods, and algorithms. John Wiley & Sons, 2011.

See also

getHDmembers