Convert orthus samples of Eta Lambda and Sigma to tidy format

Combines them all into a single tibble, see example for formatting and column headers. Primarily designed to be used by summary.orthusfit.

orthus_tidy_samples(m, use_names = FALSE, as_factor = FALSE)

Arguments

m
an object of class orthusfit

use_names
should dimension indices be replaced by dimension names if provided in data used to fit pibble model.

as_factor
if use_names should names be returned as factor?

Value

tibble

Examples

sim <- orthus_sim()
fit <- orthus(sim$Y, sim$Z, sim$X)
fit_tidy <- orthus_tidy_samples(fit, use_names=TRUE)
head(fit_tidy)

#>   Parameter       coord sample iter        val covariate coord2
#> 1       Eta log(c1/c10)     s1    1  0.1039360      <NA>   <NA>
#> 2       Eta log(c2/c10)     s1    1  2.9347975      <NA>   <NA>
#> 3       Eta log(c3/c10)     s1    1  0.3647139      <NA>   <NA>
#> 4       Eta log(c4/c10)     s1    1  1.5463812      <NA>   <NA>
#> 5       Eta log(c5/c10)     s1    1 -0.8928473      <NA>   <NA>
#> 6       Eta log(c6/c10)     s1    1  2.1186919      <NA>   <NA>