Predict response from new data

Predict response from new data

# S3 method for pibblefit
predict(object, newdata = NULL,
  response = "LambdaX", size = NULL, use_names = TRUE,
  summary = FALSE, iter = NULL, from_scratch = FALSE, ...)

Arguments

object
An object of class pibblefit

newdata
An optional matrix for which to evaluate predictions. If NULL (default), the original data of the model is used.

response
Options = "LambdaX":Mean of regression, "Eta", "Y": counts

size
the number of counts per sample if response="Y" (as vector or matrix), default if newdata=NULL and response="Y" is to use colsums of m$Y. Otherwise uses median colsums of m$Y as default. If passed as a matrix should have dimensions ncol(newdata) x iter.

use_names
if TRUE apply names to output

summary
if TRUE, posterior summary of predictions are returned rather than samples

iter
number of iterations to return if NULL uses object$iter

from_scratch
should predictions of Y come from fitted Eta or from predictions of Eta from posterior of Lambda? (default: false)

...
other arguments passed to summarise_posterior

Value

(if summary==FALSE) array D x N x iter; (if summary==TRUE) tibble with calculated posterior summaries

Details

currently only implemented for pibblefit objects in coord_system "default" "alr", or "ilr".

Examples

sim <- pibble_sim()
fit <- pibble(sim$Y, sim$X)
predict(fit)[,,1:2] # just show 2 samples

#> , , 1
#> 
#>                     s1         s2          s3         s4         s5         s6
#> log(c1/c10)  2.2072779  2.3165942  2.25948546  2.1411239  2.5310344  2.1823180
#> log(c2/c10) -0.5110173 -1.6450420 -1.05260817  0.1752501 -3.8696006 -0.2520893
#> log(c3/c10)  0.8041376  0.8828250  0.84171737  0.7565190  1.0371821  0.7861712
#> log(c4/c10) -1.2622791  0.2128430 -0.55778593 -2.1549653  3.1065149 -1.5990887
#> log(c5/c10)  5.3513516  3.6098633  4.51964645  6.4052321  0.1936748  5.7489795
#> log(c6/c10) -1.4638064 -1.1266735 -1.30279747 -1.6678260 -0.4653371 -1.5407828
#> log(c7/c10)  3.5737794  1.8848971  2.76719796  4.5958249 -1.4280972  3.9593960
#> log(c8/c10)  0.1364258 -0.1269998  0.01061819  0.2958406 -0.6437484  0.1965728
#> log(c9/c10)  3.0706228  3.6622181  3.35315860  2.7126124  4.8227204  2.9355458
#>                      s7         s8         s9        s10         s11
#> log(c1/c10)  2.30206273  2.1720521  2.4142221  2.3125514  2.25164919
#> log(c2/c10) -1.49429593 -0.1455930 -2.6578140 -1.6031029 -0.97131634
#> log(c3/c10)  0.87236509  0.7787816  0.9530990  0.8799150  0.83607671
#> log(c4/c10)  0.01675478 -1.7376174  1.5302414  0.1582892 -0.66352911
#> log(c5/c10)  3.84135966  5.9125227  2.0545794  3.6742680  4.64448391
#> log(c6/c10) -1.17148864 -1.5724429 -0.8255877 -1.1391415 -1.32696463
#> log(c7/c10)  2.10940047  4.1179990  0.3765943  1.9473563  2.88826440
#> log(c8/c10) -0.09198262  0.2213111 -0.3622593 -0.1172577  0.02950169
#> log(c9/c10)  3.58357721  2.8799891  4.1905585  3.6403394  3.31075047
#>                     s12        s13        s14        s15        s16        s17
#> log(c1/c10)  2.28719662  2.0940808  2.1999155  2.4004730  2.1333532  2.3956498
#> log(c2/c10) -1.34007794  0.6632653 -0.4346422 -2.5151841  0.2558614 -2.4651486
#> log(c3/c10)  0.86166425  0.7226568  0.7988381  0.9432022  0.7509256  0.9397304
#> log(c4/c10) -0.18384963 -2.7897680 -1.3616267  1.3447105 -2.2598232  1.2796251
#> log(c5/c10)  4.07818765  7.1546625  5.4686385  2.2736120  6.5290243  2.3504501
#> log(c6/c10) -1.21733592 -1.8129075 -1.4865118 -0.8679900 -1.6917908 -0.8828650
#> log(c7/c10)  2.33907450  5.3226169  3.6875234  0.5890105  4.7158777  0.6635274
#> log(c8/c10) -0.05615891  0.4092030  0.1541671 -0.3291274  0.3145660 -0.3175045
#> log(c9/c10)  3.50312514  2.4580259  3.0307796  4.1161517  2.6705594  4.0900493
#>                     s18        s19        s20        s21        s22         s23
#> log(c1/c10)  2.44499714  2.1740998  2.0322020  2.3172776  2.4839330  2.27473717
#> log(c2/c10) -2.97706770 -0.1668354  1.3051828 -1.6521313 -3.3809798 -1.21082621
#> log(c3/c10)  0.97525129  0.7802556  0.6781155  0.8833169  1.0032778  0.85269576
#> log(c4/c10)  1.94552171 -1.7099856 -3.6247645  0.2220646  2.4709244 -0.35197832
#> log(c5/c10)  1.56431102  5.8799014  8.1404361  3.5989766  0.9440352  4.27667571
#> log(c6/c10) -0.73067718 -1.5661278 -2.0037424 -1.1245660 -0.6105986 -1.25576102
#> log(c7/c10) -0.09886436  4.0863631  6.2786129  1.8743392 -0.7004032  2.53156674
#> log(c8/c10) -0.43641957  0.2163767  0.5583157 -0.1286466 -0.5302454 -0.02613469
#> log(c9/c10)  4.35710601  2.8910708  2.1231519  3.6659164  4.5678179  3.43569740
#>                     s24        s25        s26        s27        s28        s29
#> log(c1/c10)  2.30652445  2.0604812  2.3906931  2.3431926  2.1101063  2.1968602
#> log(c2/c10) -1.54058081  1.0118212 -2.4137290 -1.9209689  0.4970206 -0.4029471
#> log(c3/c10)  0.87557669  0.6984712  0.9361625  0.9019710  0.7341921  0.7966388
#> log(c4/c10)  0.07696145 -3.2431642  1.2127393  0.5717644 -2.5735195 -1.4028551
#> log(c5/c10)  3.77028134  7.6899294  2.4294135  3.1861307  6.8993654  5.5173116
#> log(c6/c10) -1.15772866 -1.9165292 -0.8981515 -1.0446436 -1.7634848 -1.4959344
#> log(c7/c10)  2.04046925  5.8417148  0.7401056  1.4739643  5.0750317  3.7347262
#> log(c8/c10) -0.10273425  0.4901699 -0.3055601 -0.1910956  0.3705855  0.1615297
#> log(c9/c10)  3.60772299  2.2761921  4.0632248  3.8061629  2.5447520  3.0142450
#>                     s30
#> log(c1/c10)  2.28546325
#> log(c2/c10) -1.32209635
#> log(c3/c10)  0.86041655
#> log(c4/c10) -0.20723981
#> log(c5/c10)  4.10580145
#> log(c6/c10) -1.22268165
#> log(c7/c10)  2.36585415
#> log(c8/c10) -0.05198192
#> log(c9/c10)  3.49374454
#> 
#> , , 2
#> 
#>                      s1          s2          s3          s4          s5
#> log(c1/c10)  3.24443667  3.28579751  3.26418990  3.21940671  3.36693296
#> log(c2/c10) -0.03434829 -1.12228922 -0.55393035  0.62403098 -3.25644744
#> log(c3/c10)  0.36388149  0.26843806  0.31829933  0.42164011  0.08121157
#> log(c4/c10) -1.45677459  0.28740704 -0.62378315 -2.51228504  3.70887914
#> log(c5/c10)  4.61859389  2.86955300  3.78328175  5.67704499 -0.56145130
#> log(c6/c10) -1.07058791 -0.90898405 -0.99340866 -1.16838423 -0.59197402
#> log(c7/c10)  3.75584693  1.85476918  2.84792458  4.90630478 -1.87447807
#> log(c8/c10) -0.03304931 -0.04307877 -0.03783921 -0.02697988 -0.06275304
#> log(c9/c10)  2.76056879  3.23821862  2.98868621  2.47151381  4.17519992
#>                      s6          s7          s8          s9         s10
#> log(c1/c10)  3.23499289  3.28029941  3.23110869  3.32273595  3.28426788
#> log(c2/c10)  0.21405750 -0.97766906  0.31622606 -2.09390480 -1.08205438
#> log(c3/c10)  0.38567375  0.28112537  0.39463685  0.18319969  0.27196780
#> log(c4/c10) -1.85501755  0.05555269 -2.01881368  1.84509641  0.22290274
#> log(c5/c10)  5.01794636  3.10205329  5.18219882  1.30752391  2.93423700
#> log(c6/c10) -1.10748636 -0.93046608 -1.12266258 -0.76465929 -0.91496058
#> log(c7/c10)  4.18991345  2.10747975  4.36844370  0.15695940  1.92507589
#> log(c8/c10) -0.03075932 -0.04174555 -0.02981745 -0.05203585 -0.04270785
#> log(c9/c10)  2.65150865  3.17472455  2.60665255  3.66479694  3.22055391
#>                     s11         s12         s13         s14         s15
#> log(c1/c10)  3.26122498  3.27467467  3.20160752  3.24165107  3.31753386
#> log(c2/c10) -0.47594201 -0.82971809  1.09221447  0.03892308 -1.95707092
#> log(c3/c10)  0.32514113  0.29410489  0.46271314  0.37030947  0.19520392
#> log(c4/c10) -0.74881369 -0.18164159 -3.26287455 -1.57424290  1.62572501
#> log(c5/c10)  3.90866062  3.33990838  6.42972563  4.73638947  1.52750648
#> log(c6/c10) -1.00499313 -0.95244287 -1.23792868 -1.08147171 -0.78498473
#> log(c7/c10)  2.98420211  2.36601060  5.72441283  3.88388199  0.39606414
#> log(c8/c10) -0.03712025 -0.04038163 -0.02266381 -0.03237384 -0.05077441
#> log(c9/c10)  2.95444619  3.10976812  2.26596242  2.72839971  3.60472136
#>                     s16         s17         s18         s19         s20
#> log(c1/c10)  3.21646660  3.31570893  3.33437996  3.23188346  3.17819511
#> log(c2/c10)  0.70136637 -1.90906875 -2.40018481  0.29584688  1.70804609
#> log(c3/c10)  0.42842463  0.19941508  0.15633020  0.39284901  0.51673913
#> log(c4/c10) -2.63626876  1.54876816  2.33612305 -1.98614188 -4.25017280
#> log(c5/c10)  5.80137412  1.60467773  0.81512928  5.14943599  7.41977449
#> log(c6/c10) -1.17987171 -0.79211502 -0.71916414 -1.11963543 -1.32940494
#> log(c7/c10)  5.04144133  0.47994356 -0.37823705  4.33283293  6.80052252
#> log(c8/c10) -0.02626694 -0.05033189 -0.05485936 -0.03000532 -0.01698661
#> log(c9/c10)  2.43756046  3.58364648  3.79926618  2.61559983  1.99558757
#>                     s21         s22         s23        s24         s25
#> log(c1/c10)  3.28605607  3.34911169  3.26996053  3.2819875  3.18889478
#> log(c2/c10) -1.12909040 -2.78768294 -0.70571881 -1.0220730  1.42660594
#> log(c3/c10)  0.26784140  0.12233557  0.30498316  0.2772299  0.49204880
#> log(c4/c10)  0.29831065  2.95735817 -0.38043663  0.1267409 -3.79896932
#> log(c5/c10)  2.85861901  0.19216340  3.53925727  3.0306667  6.96731396
#> log(c6/c10) -0.90797380 -0.66160478 -0.97086185 -0.9238703 -1.28759953
#> log(c7/c10)  1.84288474 -1.05535477  2.58268806  2.0298878  6.30873148
#> log(c8/c10) -0.04314146 -0.05843161 -0.03923851 -0.0421549 -0.01958114
#> log(c9/c10)  3.24120461  3.96939345  3.05532745  3.1942197  2.11915112
#>                     s26        s27        s28         s29         s30
#> log(c1/c10)  3.31383353  3.2958613  3.2076709  3.24049507  3.27401884
#> log(c2/c10) -1.85973877 -1.3870031  0.9327255  0.06933014 -0.81246722
#> log(c3/c10)  0.20374272  0.2452151  0.4487214  0.37297704  0.29561828
#> log(c4/c10)  1.46968257  0.7117950 -3.0071827 -1.62299135 -0.20929809
#> log(c5/c10)  1.68398366  2.4439826  6.1733214  4.78527373  3.36764193
#> log(c6/c10) -0.79944255 -0.8696632 -1.2142380 -1.08598841 -0.95500533
#> log(c7/c10)  0.56614322  1.3922058  5.4457204  3.93701556  2.39615492
#> log(c8/c10) -0.04987713 -0.0455191 -0.0241341 -0.03209352 -0.04022259
#> log(c9/c10)  3.56198863  3.3544387  2.3359845  2.71504979  3.10219430
#>