Provide random initialization for pibble model

Randomly initializes based on ALR transform of counts plus random pseudocounts uniformily distributed between 0 and 1.

random_pibble_init(Y)

Arguments

Y
matrix (D x N) of counts

Value

(D-1) x N matrix

Details

Notation: N is number of samples and D is number of multinomial categories

Examples

Y <- matrix(sample(1:100, 100), 10, 10)
random_pibble_init(Y)

#>             [,1]       [,2]        [,3]        [,4]       [,5]       [,6]
#>  [1,] -2.4593937 -1.8571315 -0.28109670 -0.03983349 -0.6628104 -1.2616049
#>  [2,]  1.4288773  0.7154713 -0.78445611  0.23206164 -0.5707253  0.1519485
#>  [3,]  1.3822827 -0.9342476 -0.22627389  0.19476500 -1.9329628  0.2380116
#>  [4,]  0.1175994  0.8552719  0.03241232 -1.00693347 -0.9564950  0.3039879
#>  [5,]  1.4044324  0.8771227 -2.21787727 -0.45763085 -1.3419321 -2.9820702
#>  [6,]  0.4270334  0.3793254 -0.08021703  0.08950484 -1.4449987 -0.5554563
#>  [7,]  1.0421021  0.5181841 -0.47814257 -0.25367120 -0.7236958 -0.5914296
#>  [8,]  0.6034101 -0.2247340  0.06268912 -1.16784025 -2.0738376 -0.4975917
#>  [9,]  1.1015645  0.5806827  0.04453112 -1.59611784 -0.1203932  0.2646745
#>             [,7]       [,8]        [,9]       [,10]
#>  [1,]  1.3076918  0.2437137 -0.75029403  0.81148312
#>  [2,]  1.5303625  0.5204841 -0.62170143 -1.44886207
#>  [3,]  2.4705586  0.1891045  0.29819603  1.07854391
#>  [4,]  2.1681643  0.4624971 -0.12493212  0.57867359
#>  [5,]  0.7870659 -2.1503985 -0.02842734 -0.06138690
#>  [6,] -1.0527227 -1.1429545  0.42445271 -0.57309446
#>  [7,]  1.6706995 -0.4552085 -1.40440082  0.72575562
#>  [8,] -0.4998857  0.5826457 -1.51459606  0.08984932
#>  [9,]  1.7331345  0.2628816  0.15644365  0.66024669