Sample from the prior distribution of pibblefit object

Note this can be used to sample from prior and then predict can be called to get counts or LambdaX (predict.pibblefit)

# S3 method for pibblefit
sample_prior(m, n_samples = 2000, pars = c("Eta",
  "Lambda", "Sigma"), use_names = TRUE, ...)

Arguments

m
object of class pibblefit

n_samples
number of samples to produce

pars
parameters to sample

use_names
should names be used if available

...
currently ignored

Details

Could be greatly speed up in the future if needed by sampling directly from cholesky form of inverse wishart (currently implemented as header in this library - see MatDist.h).

Examples

# Sample prior of already fitted  pibblefit object
sim <- pibble_sim()
attach(sim)
fit <- pibble(Y, X)
sample_prior(fit)

#>  pibblefit Object (Priors Only): 
#>   Number of Samples:		 30 
#>   Number of Categories:		 10 
#>   Number of Covariates:		 2 
#>   Number of Posterior Samples:	 2000 
#>   Contains Samples of Parameters:Eta  Lambda  Sigma
#>   Coordinate System:		 alr, reference category: 10 [c10] 

# Sample prior as part of model fitting
m <- pibblefit(N=as.integer(sim$N), D=as.integer(sim$D), Q=as.integer(sim$Q),
                iter=2000, upsilon=upsilon,
                Xi=Xi, Gamma=Gamma, Theta=Theta, X=X,
                coord_system="alr", alr_base=D)

#> Error in eval(quote(y)): is.integer(m$iter) is not TRUE

m <- sample_prior(pibblefit)

#> Error in UseMethod("sample_prior", m): no applicable method for 'sample_prior' applied to an object of class "function"

plot(m) # plot prior distribution (defaults to parameter Lambda) 

#> Error in plot(m): object 'm' not found