Transform Fit Stray Parameters to other representations

These are a collection of convenience functions for transforming stray fit objects to a number of different representations including ILR bases, CLR coordinates, ALR coordinates, and proportions.

to_proportions(m)

to_alr(m, d)

to_ilr(m, V = NULL)

to_clr(m)

Arguments

m
object of class pibblefit (e.g., output of pibble)

d
(integer) multinomial category to take as new alr reference

V
(matrix) contrast matrix for ILR basis to transform into to (defaults to driver::create_default_ilr_base(D))

Value

pibblefit object

Details

Note: that there is a degeneracy of representations for a covariance matrix represented in terms of proportions. As such the function to_proportions does not attempt to transform parameters Sigma or prior Xi and instead just removes them from the pibblefit object returned.

Examples

# NOT RUN {
m <- pibble(Y, X)
m.prop <- to_proportions(m)
# convert back to default coordinates (alr with D-th part as reference)
m <- to_alr(m.prop, ncategories(m))
V <- driver::create_default_ilr_base(ncategories(m))
m.ilr <- to_ilr(m, V)
m.clr <- to_clr(m)
# }