Transform Lambda into IQLR (Inter-Quantile Log-Ratio)

Takes idea from Wu et al. (citation below) and calculates IQLR for Lambda, potentially useful if you believe there is an invariant group of categories (e.g., taxa / genes) that are not changing (in absolute abundance) between samples. IQLR is defined as $$IQLR_x = log(x_i/g(IQVF))$$ for i in 1,...,D. IQVF are the CLR coordinates whose variance is within the inter-quantile range (defined by probs argument to this function). A different IQVF is fit for each posteior sample as the IQVFs are calculted based on posterior estimates for Lambda. The variance of a CLR coordinate is defined as the norm of each row of Lambda[,focus.cov] (i.e., the covariation in Eta, explained by those covariates). This definition of variance allows uses to exclude variation from technical / trivial sources in calculation of IQVF/IQLR.

lambda_to_iqlr(m, focus.cov, probs = c(0.25, 0.75))

Arguments

m
object of class pibblefit (e.g., output of pibble)

focus.cov
vector of integers or characters specifying columns (covariates) of Lambda to include in calculating IQLR

probs
bounds for categories (i.e., features / genes / taxa) to include in calculation of iqlr (smaller bounds means more stringent inclusion criteria)

Value

array of dimension (D, Q, iter) where D is number of taxa, Q is number of covariates, and iter is number of posterior samples.

Details

Primarily intended for doing differential expression analysis under assumption that only small group of categories (e.g., taxa / genes) are changing

References

Jia R. Wu, Jean M. Macklaim, Briana L. Genge, Gregory B. Gloor (2017) Finding the center: corrections for asymmetry in high-throughput sequencing datasets. arxiv:1704.01841v1

Examples

sim <- pibble_sim()
fit <- pibble(sim$Y, sim$X)
# Use first two covariates to define iqlr, just show first 5 samples
lambda_to_iqlr(fit, 1:2)[,,1:5]

#> , , 1
#> 
#>              [,1]        [,2]
#>  [1,]  0.64096613 -0.14983658
#>  [2,] -0.35708899  0.10418474
#>  [3,]  0.39387530  0.41716424
#>  [4,] -0.02193774 -0.38645200
#>  [5,]  1.52424184  0.57650978
#>  [6,]  0.18756434 -0.56359596
#>  [7,]  0.07387803 -0.27420062
#>  [8,]  0.81964609 -1.84881610
#>  [9,] -0.52665841  1.88072060
#> [10,] -0.22435065  0.04224698
#> 
#> , , 2
#> 
#>             [,1]       [,2]
#>  [1,]  0.5823358  0.4835876
#>  [2,] -0.1484976  0.3130282
#>  [3,]  0.2791348  0.4492592
#>  [4,] -0.2547597  0.3372564
#>  [5,]  2.0944634  0.7993226
#>  [6,]  0.8033246 -0.9848727
#>  [7,]  0.3962920 -0.2386476
#>  [8,]  0.2438678 -1.9162546
#>  [9,] -0.2240591  1.1516934
#> [10,] -0.4000673  0.3345882
#> 
#> , , 3
#> 
#>              [,1]       [,2]
#>  [1,]  0.42566528  0.2590140
#>  [2,] -0.12997334  0.2559298
#>  [3,]  0.21650786 -0.1673776
#>  [4,]  0.21004296 -0.1734227
#>  [5,]  0.38489965  1.1633263
#>  [6,]  0.25976842 -0.1569353
#>  [7,]  0.08674747  0.3972437
#>  [8,] -0.92179516 -0.9033046
#>  [9,] -0.50144222  1.5095633
#> [10,] -0.68631924  0.4977356
#> 
#> , , 4
#> 
#>               [,1]        [,2]
#>  [1,]  0.756653632  0.38114995
#>  [2,] -0.161496585  0.35759198
#>  [3,]  0.244348993 -0.08666121
#>  [4,] -0.650185919 -0.06353822
#>  [5,]  1.794226006  1.25050924
#>  [6,]  0.101849295 -0.70513621
#>  [7,]  0.009435609  0.51419823
#>  [8,]  0.508249424 -1.31540548
#>  [9,]  0.159401537  1.70352204
#> [10,] -0.208317009  0.38752448
#> 
#> , , 5
#> 
#>               [,1]       [,2]
#>  [1,]  0.385035482  0.8271314
#>  [2,] -0.517243497  0.7773113
#>  [3,]  0.580085165 -0.3603764
#>  [4,] -0.496769348 -0.1102662
#>  [5,]  1.165490924  1.3219374
#>  [6,]  0.155936427 -0.8414836
#>  [7,] -0.239252244  1.3121262
#>  [8,]  0.327688133 -1.1957472
#>  [9,] -0.008179498  2.0326680
#> [10,] -0.633149146  0.7820630
#>