Calculations for the Collapsed Pibble Model

Functions providing access to the Log Likelihood, Gradient, and Hessian of the collapsed pibble model. Note: These are convenience functions but are not as optimized as direct coding of the PibbleCollapsed C++ class due to a lack of Memoization. By contrast function optimPibbleCollapsed is much more optimized and massively cuts down on repeated calculations. A more efficient Rcpp module based implementation of these functions may following if the future. For model details see optimPibbleCollapsed documentation

loglikPibbleCollapsed(Y, upsilon, ThetaX, K, A, eta, sylv = FALSE)

gradPibbleCollapsed(Y, upsilon, ThetaX, K, A, eta, sylv = FALSE)

hessPibbleCollapsed(Y, upsilon, ThetaX, K, A, eta, sylv = FALSE)

Arguments

Y
D x N matrix of counts

upsilon
(must be > D)

ThetaX
D-1 x N matrix formed by Theta*X (Theta is Prior mean for regression coefficients)

eta
matrix (D-1)xN of parameter values at which to calculate quantities

sylv
(default:false) if true and if N < D-1 will use sylvester determinant identity to speed computation

Value

see below

loglikPibbleCollapsed - double

gradPibbleCollapsed - vector

hessPibbleCollapsed- matrix

Examples

D <- 10
Q <- 2
N <- 30

# Simulate Data
Sigma <- diag(sample(1:8, D-1, replace=TRUE))
Sigma[2, 3] <- Sigma[3,2] <- -1
Gamma <- diag(sqrt(rnorm(Q)^2))
Theta <- matrix(0, D-1, Q)
Phi <-  Theta + t(chol(Sigma))%*%matrix(rnorm(Q*(D-1)), nrow=D-1)%*%chol(Gamma)
X <- matrix(rnorm(N*(Q-1)), Q-1, N)
X <- rbind(1, X)
Eta <- Phi%*%X + t(chol(Sigma))%*%matrix(rnorm(N*(D-1)), nrow=D-1)
Pi <- t(driver::alrInv(t(Eta)))
Y <- matrix(0, D, N)
for (i in 1:N) Y[,i] <- rmultinom(1, sample(5000:10000), prob = Pi[,i])

# Priors
upsilon <- D+10
Xi <- Sigma*(upsilon-D)

# Precompute
K <- solve(Xi)
A <- solve(diag(N)+ t(X)%*%Gamma%*%X)
ThetaX <- Theta%*%X

loglikPibbleCollapsed(Y, upsilon, ThetaX, K, A, Eta)

#> [1] -120985

gradPibbleCollapsed(Y, upsilon, ThetaX, K, A, Eta)[1:5]

#> [1] -0.2919848 -0.1266784  1.1792351 13.6813417 -3.7704904

hessPibbleCollapsed(Y, upsilon, ThetaX, K, A, Eta)[1:5,1:5]

#>             [,1]        [,2]        [,3]          [,4]         [,5]
#> [1,] -0.57310877 -0.04991180  0.03243245     0.2647870  -0.01069384
#> [2,] -0.04991180 -0.66136243 -0.07135975     0.2429493  -0.02165481
#> [3,]  0.03243245 -0.07135975 -1.13581834     0.8367218  -0.02528917
#> [4,]  0.26478698  0.24294926  0.83672183 -1157.3227611  17.52554779
#> [5,] -0.01069384 -0.02165481 -0.02528917    17.5255478 -21.25162872