Predict response from new data

Predict response from new data

# S3 method for pibblefit
predict(
  object,
  newdata = NULL,
  response = "LambdaX",
  size = NULL,
  use_names = TRUE,
  summary = FALSE,
  iter = NULL,
  from_scratch = FALSE,
  ...
)

Arguments

object
An object of class pibblefit

newdata
An optional matrix for which to evaluate predictions. If NULL (default), the original data of the model is used.

response
Options = "LambdaX":Mean of regression, "Eta", "Y": counts

size
the number of counts per sample if response="Y" (as vector or matrix), default if newdata=NULL and response="Y" is to use colsums of m$Y. Otherwise uses median colsums of m$Y as default. If passed as a matrix should have dimensions ncol(newdata) x iter.

use_names
if TRUE apply names to output

summary
if TRUE, posterior summary of predictions are returned rather than samples

iter
number of iterations to return if NULL uses object$iter

from_scratch
should predictions of Y come from fitted Eta or from predictions of Eta from posterior of Lambda? (default: false)

...
other arguments passed to summarise_posterior

Value

(if summary==FALSE) array D x N x iter; (if summary==TRUE) tibble with calculated posterior summaries

Details

currently only implemented for pibblefit objects in coord_system "default" "alr", or "ilr".

Examples

sim <- pibble_sim()
fit <- pibble(sim$Y, sim$X)
predict(fit)[,,1:2] # just show 2 samples

#> , , 1
#> 
#>                      s1         s2           s3         s4         s5
#> log(c1/c10) -4.64996425 -6.6758735 -4.142084857  0.0751788 -1.9244730
#> log(c2/c10)  0.05927058 -0.2727557  0.142506936  0.8336743  0.5059513
#> log(c3/c10) -0.04979552 -0.2396367 -0.002203857  0.3929817  0.2056011
#> log(c4/c10)  3.34811840  4.5236093  3.053432145  0.6064543  1.7667099
#> log(c5/c10) -2.48114499 -3.4465012 -2.239137838 -0.2295899 -1.1824343
#> log(c6/c10)  2.75705540  5.5473777  2.057543701 -3.7509717 -0.9968142
#> log(c7/c10)  6.19857905  8.4281842  5.639634703  0.9983444  3.1990521
#> log(c8/c10) -3.64599529 -5.3573501 -3.216972190  0.3454947 -1.3436796
#> log(c9/c10)  0.19128891  0.9736005 -0.004830413 -1.6333408 -0.8611686
#>                     s6           s7         s8         s9        s10
#> log(c1/c10)  1.4679153 -5.003850629 -0.8002855 -0.5848094 -0.1730725
#> log(c2/c10)  1.0619298  0.001272147  0.6901944  0.7255088  0.7929884
#> log(c3/c10)  0.5234904 -0.082957025  0.3109449  0.3311364  0.3697189
#> log(c4/c10) -0.2016516  3.553453477  1.1144239  0.9893984  0.7504968
#> log(c5/c10)  0.4340562 -2.649773677 -0.6467532 -0.5440777 -0.3478830
#> log(c6/c10) -5.6692135  3.244469666 -2.5451785 -2.8419576 -3.4090505
#> log(c7/c10) -0.5344254  6.588047091  1.9618327  1.7246915  1.2715563
#> log(c8/c10)  1.5219869 -3.944935226 -0.3940399 -0.2120199  0.1357884
#> log(c9/c10) -2.1711507  0.327943314 -1.2952774 -1.3784842 -1.5374778
#>                     s11         s12        s13        s14         s15
#> log(c1/c10) -4.52139879 -4.73071532 -5.7188997 -2.9858327 -3.75671401
#> log(c2/c10)  0.08034118  0.04603629 -0.1159172  0.3320051  0.20566536
#> log(c3/c10) -0.03774809 -0.05736244 -0.1499619  0.1061446  0.03390795
#> log(c4/c10)  3.27352102  3.39497250  3.9683455  2.3825414  2.82982888
#> log(c5/c10) -2.41988288 -2.51962329 -2.9904982 -1.6881777 -2.05550664
#> log(c6/c10)  2.57997980  2.86827535  4.2293200  0.4650162  1.52676529
#> log(c7/c10)  6.05708692  6.28744928  7.3749911  4.3671267  5.21551657
#> log(c8/c10) -3.53739165 -3.71420849 -4.5489616 -2.2402465 -2.89143625
#> log(c9/c10)  0.14164293  0.22247120  0.6040619 -0.4513210 -0.15364265
#>                     s16         s17        s18        s19         s20
#> log(c1/c10) -4.66017251 -4.45044030 -2.5137116 -2.1867602 -3.57927408
#> log(c2/c10)  0.05759755  0.09197056  0.4093810  0.4629651  0.23474599
#> log(c3/c10) -0.05075211 -0.03109881  0.1503855  0.1810230  0.05053524
#> log(c4/c10)  3.35404153  3.23234886  2.1086031  1.9188964  2.72687311
#> log(c5/c10) -2.48600928 -2.38607080 -1.4632096 -1.3074155 -1.97095560
#> log(c6/c10)  2.77111543  2.48224736 -0.1852449 -0.6355612  1.28237398
#> log(c7/c10)  6.20981371  5.97899388  3.8475360  3.4877111  5.02023587
#> log(c8/c10) -3.65461856 -3.47745058 -1.8414296 -1.5652425 -2.74154667
#> log(c9/c10)  0.19523086  0.11424207 -0.6336322 -0.7598856 -0.22216167
#>                      s21         s22         s23         s24        s25
#> log(c1/c10) -4.197346799 -3.65798857 -4.30143027 -0.35985959 -1.4917090
#> log(c2/c10)  0.133450056  0.22184547  0.11639182  0.76237582  0.5768770
#> log(c3/c10) -0.007382268  0.04315917 -0.01713558  0.35221573  0.2461539
#> log(c4/c10)  3.085496716  2.77254553  3.14588895  0.85887606  1.5156077
#> log(c5/c10) -2.265470438 -2.00846346 -2.31506675 -0.43688797 -0.9762201
#> log(c6/c10)  2.133656997  1.39078892  2.27701309 -3.15178524 -1.5928681
#> log(c7/c10)  5.700452981  5.10686475  5.81500157  1.47712394  2.7227756
#> log(c8/c10) -3.263653844 -2.80803950 -3.35157671 -0.02199709 -0.9781090
#> log(c9/c10)  0.016509171 -0.19176581  0.05670135 -1.46534931 -1.0282819
#>                     s26         s27        s28        s29        s30
#> log(c1/c10) -4.47563060 -3.31232995 -0.8649067 -0.2120651 -7.3769854
#> log(c2/c10)  0.08784213  0.27849546  0.6796036  0.7865979 -0.3876609
#> log(c3/c10) -0.03345930  0.07554968  0.3048895  0.3660651 -0.3053355
#> log(c4/c10)  3.24696499  2.57198444  1.1519190  0.7731215  4.9304147
#> log(c5/c10) -2.39807410 -1.84375534 -0.6775456 -0.3664632 -3.7805847
#> log(c6/c10)  2.51694243  0.91470689 -2.4561744 -3.3553453  6.5130322
#> log(c7/c10)  6.00671694  4.72645172  2.0329513  1.3144694  9.1997898
#> log(c8/c10) -3.49872969 -2.51604983 -0.4486277  0.1028499 -5.9496034
#> log(c9/c10)  0.12396939 -0.32524304 -1.2703237 -1.5224206  1.2443372
#> 
#> , , 2
#> 
#>                     s1         s2         s3          s4         s5         s6
#> log(c1/c10) -3.1537620 -4.8415084 -2.7306574  0.78266460 -0.8832072  1.9429268
#> log(c2/c10)  0.7034841  0.8405557  0.6691214  0.38378440  0.5190794  0.2895529
#> log(c3/c10) -0.3670812 -0.8531123 -0.2452371  0.76651666  0.2867849  1.1006448
#> log(c4/c10)  2.5130808  3.2678591  2.3238638  0.75266841  1.4976642  0.2337866
#> log(c5/c10) -1.5632350 -1.8582597 -1.4892747 -0.87513238 -1.1663332 -0.6723140
#> log(c6/c10)  1.9438358  3.7871813  1.4817237 -2.35550343 -0.5360492 -3.6227342
#> log(c7/c10)  5.8535404  8.4059019  5.2136835 -0.09947722  2.4198036 -1.8541299
#> log(c8/c10) -3.2245485 -4.4093201 -2.9275357 -0.46123855 -1.6306545  0.3532474
#> log(c9/c10)  0.4047366  1.5952113  0.1062940 -2.37187505 -1.1968299 -3.1902817
#>                     s7          s8         s9        s10        s11        s12
#> log(c1/c10) -3.4485780  0.05333199  0.2328410  0.5758512 -3.0466566 -3.2210342
#> log(c2/c10)  0.7274278  0.44301767  0.4284387  0.4005809  0.6947855  0.7089477
#> log(c3/c10) -0.4519813  0.55648609  0.6081804  0.7069593 -0.3362374 -0.3864540
#> log(c4/c10)  2.6449257  1.07883381  0.9985555  0.8451576  2.4651821  2.5431656
#> log(c5/c10) -1.6147700 -1.00262255 -0.9712437 -0.9112842 -1.5445126 -1.5749945
#> log(c6/c10)  2.2658319 -1.55893114 -1.7549897 -2.1296232  1.8268560  2.0173101
#> log(c7/c10)  6.2993876  1.00348501  0.7320154  0.2132846  5.6915659  5.9552754
#> log(c8/c10) -3.4315048 -0.97321866 -0.8472061 -0.6064183 -3.1493622 -3.2717726
#> log(c9/c10)  0.6126891 -1.85743044 -1.9840495 -2.2259964  0.3291883  0.4521879
#>                    s13         s14        s15        s16        s17
#> log(c1/c10) -4.0442717 -1.76740576 -2.4096122 -3.1622663 -2.9875424
#> log(c2/c10)  0.7758075  0.59089020  0.6430475  0.7041748  0.6899845
#> log(c3/c10) -0.6235270  0.03215663 -0.1527837 -0.3695303 -0.3192139
#> log(c4/c10)  2.9113263  1.89308734  2.1802890  2.5168840  2.4387456
#> log(c5/c10) -1.7188996 -1.32089459 -1.4331548 -1.5647216 -1.5341792
#> log(c6/c10)  2.9164447  0.42966666  1.1310804  1.9531242  1.7622919
#> log(c7/c10)  7.2002491  3.75696804  4.7281703  5.8664013  5.6021682
#> log(c8/c10) -3.8496725 -2.25134812 -2.7021670 -3.2305184 -3.1078650
#> log(c9/c10)  1.0328697 -0.57314848 -0.1201595  0.4107352  0.2874914
#>                       s18        s19        s20        s21        s22
#> log(c1/c10) -1.374091e+00 -1.1017136 -2.2617904 -2.7766950 -2.3273660
#> log(c2/c10)  5.589468e-01  0.5368255  0.6310420  0.6728604  0.6363678
#> log(c3/c10)  1.454221e-01  0.2238602 -0.1102145 -0.2584948 -0.1290987
#> log(c4/c10)  1.717193e+00  1.5953827  2.1141815  2.3444524  2.1435076
#> log(c5/c10) -1.252142e+00 -1.2045290 -1.4073149 -1.4973223 -1.4187778
#> log(c6/c10)  9.049237e-05 -0.2973979  0.9696303  1.5320058  1.0412515
#> log(c7/c10)  3.162162e+00  2.7502487  4.5046209  5.2833058  4.6037901
#> log(c8/c10) -1.975247e+00 -1.7840427 -2.5983984 -2.9598534 -2.6444314
#> log(c9/c10) -8.505786e-01 -1.0427034 -0.2244276  0.1387673 -0.1781730
#>                    s23        s24        s25        s26         s27
#> log(c1/c10) -2.8634050  0.4202424 -0.5226797 -3.0085279 -2.03940435
#> log(c2/c10)  0.6799026  0.4132188  0.4897989  0.6916888  0.61298076
#> log(c3/c10) -0.2834652  0.6621476  0.3906083 -0.3252572 -0.04617254
#> log(c4/c10)  2.3832300  0.9147475  1.3364325  2.4481306  2.01472804
#> log(c5/c10) -1.5124795 -0.9384852 -1.1033116 -1.5378476 -1.36844100
#> log(c6/c10)  1.6267098 -1.9596683 -0.9298149  1.7852122  0.72674174
#> log(c7/c10)  5.4144364  0.4486101  1.8745816  5.6339044  4.16830873
#> log(c8/c10) -3.0207225 -0.7156533 -1.3775699 -3.1225965 -2.44228689
#> log(c9/c10)  0.1999294 -2.1162357 -1.4511328  0.3022938 -0.38129060
#>                       s28        s29        s30
#> log(c1/c10) -0.0005027419  0.5433672 -5.4255914
#> log(c2/c10)  0.4473898992  0.4032191  0.8879924
#> log(c3/c10)  0.5409829624  0.6976046 -1.0213144
#> log(c4/c10)  1.1029092820  0.8596848  3.5290673
#> log(c5/c10) -1.0120330749 -0.9169625 -1.9603597
#> log(c6/c10) -1.5001332006 -2.0941443  4.4251129
#> log(c7/c10)  1.0848987256  0.2624098  9.2892046
#> log(c8/c10) -1.0110097970 -0.6292216 -4.8193373
#> log(c9/c10) -1.8194573847 -2.2030834  2.0072021
#>