Predict response from new data

Predict response from new data

# S3 method for pibblefit
predict(object, newdata = NULL,
  response = "LambdaX", size = NULL, use_names = TRUE,
  summary = FALSE, iter = NULL, from_scratch = FALSE, ...)

Arguments

object
An object of class pibblefit

newdata
An optional matrix for which to evaluate predictions. If NULL (default), the original data of the model is used.

response
Options = "LambdaX":Mean of regression, "Eta", "Y": counts

size
the number of counts per sample if response="Y" (as vector or matrix), default if newdata=NULL and response="Y" is to use colsums of m$Y. Otherwise uses median colsums of m$Y as default. If passed as a matrix should have dimensions ncol(newdata) x iter.

use_names
if TRUE apply names to output

summary
if TRUE, posterior summary of predictions are returned rather than samples

iter
number of iterations to return if NULL uses object$iter

from_scratch
should predictions of Y come from fitted Eta or from predictions of Eta from posterior of Lambda? (default: false)

...
other arguments passed to summarise_posterior

Value

(if summary==FALSE) array D x N x iter; (if summary==TRUE) tibble with calculated posterior summaries

Details

currently only implemented for pibblefit objects in coord_system "default" "alr", or "ilr".

Examples

sim <- pibble_sim()
fit <- pibble(sim$Y, sim$X)
predict(fit)[,,1:2] # just show 2 samples

#> , , 1
#> 
#>                      s1         s2         s3           s4          s5
#> log(c1/c10)  1.14618767 -0.4613550  0.1034769  1.507931991  2.94381019
#> log(c2/c10) -2.68692122 -0.2840397 -1.1283246 -3.227640141 -5.37392544
#> log(c3/c10) -1.64838164 -1.9122642 -1.8195455 -1.589000317 -1.35329697
#> log(c4/c10) -0.01454021  3.3532715  2.1699453 -0.772396784 -3.78057045
#> log(c5/c10) -0.56365603 -3.0571677 -2.1810386 -0.002542664  2.22469467
#> log(c6/c10)  0.40487999  0.8491763  0.6930668  0.304900266 -0.09195104
#> log(c7/c10)  0.30361953  0.1224224  0.1860885  0.344394200  0.50624182
#> log(c8/c10)  1.18929288  2.7776137  2.2195357  0.831874039 -0.58683496
#> log(c9/c10)  0.80208789 -0.7874318 -0.2289326  1.159776525  2.57955643
#>                     s6         s7         s8          s9        s10        s11
#> log(c1/c10)  3.6980966  0.4442964  2.2190942  2.75169871 -0.2429051  3.5254205
#> log(c2/c10) -6.5013983 -1.6377662 -4.2906530 -5.08676595 -0.6105687 -6.2432900
#> log(c3/c10) -1.2294788 -1.7635990 -1.4722611 -1.38483260 -1.8764050 -1.2578240
#> log(c4/c10) -5.3608051  1.4559262 -2.2622861 -3.37809572  2.8956177 -4.9990476
#> log(c5/c10)  3.3946928 -1.6523823  1.1005642  1.92670431 -2.7183229  3.1268493
#> log(c6/c10) -0.3004224  0.5988703  0.1083476 -0.03885483  0.7888006 -0.2526978
#> log(c7/c10)  0.5912626  0.2245046  0.4245541  0.48458762  0.1470454  0.5717991
#> log(c8/c10) -1.3321021  1.8827914  0.1292154 -0.39702061  2.5617759 -1.1614908
#> log(c9/c10)  3.3253862  0.1080659  1.8629656  2.38959880 -0.5714311  3.1546460
#>                    s12        s13        s14         s15         s16        s17
#> log(c1/c10)  2.0328711  1.2332860  3.8678409  2.72968436 -0.07941895  0.6697704
#> log(c2/c10) -4.0122951 -2.8171119 -6.7551245 -5.05385991 -0.85494030 -1.9747944
#> log(c3/c10) -1.5028301 -1.6340842 -1.2016148 -1.38844631 -1.84956833 -1.7265868
#> log(c4/c10) -1.8721474 -0.1970118 -5.7164205 -3.33197553  2.55311320  0.9835569
#> log(c5/c10)  0.8117074 -0.4285549  3.6579888  1.89255715 -2.46473429 -1.3026423
#> log(c6/c10)  0.1598164  0.3808075 -0.3473368 -0.03277046  0.74361596  0.5365533
#> log(c7/c10)  0.4035637  0.3134370  0.6103956  0.48210623  0.16547306  0.2499193
#> log(c8/c10)  0.3132118  1.1032360 -1.4998168 -0.37526950  2.40024457  1.6600135
#> log(c9/c10)  1.6788303  0.8882098  3.4932274  2.36783127 -0.40977789  0.3310120
#>                     s18        s19        s20        s21         s22
#> log(c1/c10)  1.56092435  0.0991655  1.3685260 -0.1657248  0.02642478
#> log(c2/c10) -3.30685071 -1.1218802 -3.0192624 -0.7259343 -1.01315067
#> log(c3/c10) -1.58030148 -1.8202532 -1.6118842 -1.8637357 -1.83219379
#> log(c4/c10) -0.88341610  2.1789777 -0.4803404  2.7339245  2.33136995
#> log(c5/c10)  0.07965551 -2.1877262 -0.2187798 -2.5986061 -2.30055664
#> log(c6/c10)  0.29025412  0.6942584  0.3434296  0.7674694  0.71436263
#> log(c7/c10)  0.35036733  0.1856025  0.3286808  0.1557449  0.17740343
#> log(c8/c10)  0.77951532  2.2237955  0.9696131  2.4855184  2.29566645
#> log(c9/c10)  1.21217476 -0.2331956  1.0219335 -0.4951161 -0.30512082
#>                     s23         s24        s25        s26         s27
#> log(c1/c10) -0.09000412  2.46208437  2.0100137  0.9123498  2.44854596
#> log(c2/c10) -0.83911806 -4.65386364 -3.9781290 -2.3373910 -4.63362704
#> log(c3/c10) -1.85130591 -1.43237359 -1.5065822 -1.6867668 -1.43459595
#> log(c4/c10)  2.57528920 -2.77135189 -1.8242610  0.4753517 -2.74298884
#> log(c5/c10) -2.48115329  1.47747410  0.7762525 -0.9263696  1.45647423
#> log(c6/c10)  0.74654151  0.04118944  0.1661338  0.4695087  0.04493122
#> log(c7/c10)  0.16427993  0.45194322  0.4009872  0.2772621  0.45041721
#> log(c8/c10)  2.41070318 -0.11086927  0.3357959  1.4203347 -0.09749275
#> log(c9/c10) -0.42024438  2.10323146  1.6562291  0.5708716  2.08984484
#>                     s28        s29        s30
#> log(c1/c10)  0.06832755  1.7124108  1.3497539
#> log(c2/c10) -1.07578502 -3.5332857 -2.9912028
#> log(c3/c10) -1.82531534 -1.5554346 -1.6149657
#> log(c4/c10)  2.24358339 -1.2007810 -0.4410128
#> log(c5/c10) -2.23555990  0.3146310 -0.2478978
#> log(c6/c10)  0.70278145  0.2483860  0.3486179
#> log(c7/c10)  0.18212657  0.3674424  0.3265649
#> log(c8/c10)  2.25426473  0.6298403  0.9881607
#> log(c9/c10) -0.26368785  1.3619628  1.0033719
#> 
#> , , 2
#> 
#>                     s1          s2         s3         s4         s5         s6
#> log(c1/c10)  1.3449985  0.17771469  0.5878557  1.6076717  2.6503049  3.1980144
#> log(c2/c10) -2.7038983 -1.25154013 -1.7618458 -3.0307216 -4.3279869 -5.0094581
#> log(c3/c10) -1.8730038 -0.69918214 -1.1116203 -2.1371482 -3.1856211 -3.7363982
#> log(c4/c10) -0.7148461  2.92666635  1.6471717 -1.5342934 -4.7869402 -6.4955998
#> log(c5/c10) -0.9158204 -2.20580761 -1.7525532 -0.6255354  0.5266981  1.1319820
#> log(c6/c10)  0.5988256  2.06675253  1.5509765  0.2684989 -1.0426728 -1.7314491
#> log(c7/c10)  0.2994293  0.08156718  0.1581160  0.3484546  0.5430519  0.6452765
#> log(c8/c10)  1.2196871  3.05679799  2.4113042  0.8062832 -0.8346483 -1.6966520
#> log(c9/c10)  0.3073840 -0.61759258 -0.2925895  0.5155309  1.3417322  1.7757470
#>                      s7           s8         s9        s10        s11
#> log(c1/c10)  0.83533473  2.124067432  2.5108072  0.3363376  3.0726292
#> log(c2/c10) -2.06976429 -3.673231595 -4.1544210 -1.4489019 -4.8534514
#> log(c3/c10) -1.36048544 -2.656436194 -3.0453420 -0.8586934 -3.6103108
#> log(c4/c10)  0.87512457 -3.145265376 -4.3517565  2.4318191 -6.1044425
#> log(c5/c10) -1.47905942 -0.054856816  0.3725365 -2.0305105  0.9934165
#> log(c6/c10)  1.23975728 -0.380898657 -0.8672462  1.8672751 -1.5737700
#> log(c7/c10)  0.20430552  0.444834853  0.5170160  0.1111726  0.6218746
#> log(c8/c10)  2.02181329 -0.006437908 -0.6151020  2.8071519 -1.4993166
#> log(c9/c10) -0.09648269  0.924732181  1.2311917 -0.4918970  1.6763895
#>                    s12        s13        s14        s15        s16         s17
#> log(c1/c10)  1.9888453  1.4082432  3.3212707  2.4948219  0.4550496  0.99905799
#> log(c2/c10) -3.5049855 -2.7825886 -5.1628162 -4.1345318 -1.5966059 -2.27347209
#> log(c3/c10) -2.5204567 -1.9366027 -3.8603449 -3.0292672 -0.9780704 -1.52512570
#> log(c4/c10) -2.7234203 -0.9121471 -6.8801160 -4.3018881  2.0614795  0.36436592
#> log(c5/c10) -0.2042933 -0.8459275  1.2681949  0.3548709 -1.8993196 -1.29812577
#> log(c6/c10) -0.2108490  0.5192917 -1.8864511 -0.8471438  1.7179878  1.03386580
#> log(c7/c10)  0.4195970  0.3112333  0.6682811  0.5140325  0.1333290  0.23486286
#> log(c8/c10)  0.2063792  1.1201505 -1.8906370 -0.5899439  2.6203189  1.76414007
#> log(c9/c10)  0.8175798  0.3575003  1.8734175  1.2185247 -0.3978274  0.03325456
#>                    s18        s19        s20        s21        s22        s23
#> log(c1/c10)  1.6461510  0.5847250  1.5064449  0.3923804  0.5319059  0.4473634
#> log(c2/c10) -3.0785983 -1.7579506 -2.9047731 -1.5186316 -1.6922320 -1.5870425
#> log(c3/c10) -2.1758430 -1.1084721 -2.0353544 -0.9150502 -1.0553571 -0.9703412
#> log(c4/c10) -1.6543352  1.6569381 -1.2185017  2.2569852  1.8217152  2.0854577
#> log(c5/c10) -0.5830112 -1.7560129 -0.7374030 -1.9685765 -1.8143843 -1.9078137
#> log(c6/c10)  0.2201089  1.5549135  0.3957974  1.7967979  1.6213366  1.7276536
#> log(c7/c10)  0.3556364  0.1575317  0.3295617  0.1216324  0.1476735  0.1318945
#> log(c8/c10)  0.7457232  2.4162313  0.9655974  2.7189498  2.4993599  2.6324157
#> log(c9/c10)  0.5460226 -0.2950703  0.4353170 -0.4474876 -0.3369252 -0.4039181
#>                    s24        s25        s26        s27        s28         s29
#> log(c1/c10)  2.3005097  1.9722479  1.1752020  2.2906791  0.5623327  1.75614970
#> log(c2/c10) -3.8927646 -3.4843347 -2.4926340 -3.8805332 -1.7300896 -3.21546091
#> log(c3/c10) -2.8338667 -2.5037663 -1.7022562 -2.8239810 -1.0859543 -2.28645780
#> log(c4/c10) -3.6957028 -2.6716422 -0.1851409 -3.6650347  1.7267943 -1.99749227
#> log(c5/c10)  0.1401328 -0.2226354 -1.1034658  0.1292688 -1.7807591 -0.46144957
#> log(c6/c10) -0.6027850 -0.1899768  0.8123546 -0.5904224  1.5830731  0.08177917
#> log(c7/c10)  0.4777661  0.4164992  0.2677384  0.4759313  0.1533524  0.37616661
#> log(c8/c10) -0.2841287  0.2325008  1.4869187 -0.2686569  2.4514731  0.57260348
#> log(c9/c10)  1.0645482  0.8044277  0.1728342  1.0567582 -0.3128144  0.63318759
#>                    s30
#> log(c1/c10)  1.4928140
#> log(c2/c10) -2.8878132
#> log(c3/c10) -2.0216471
#> log(c4/c10) -1.1759779
#> log(c5/c10) -0.7524668
#> log(c6/c10)  0.4129391
#> log(c7/c10)  0.3270176
#> log(c8/c10)  0.9870502
#> log(c9/c10)  0.4245156
#>