Predict response from new data

Predict response from new data

# S3 method for pibblefit
predict(object, newdata = NULL,
  response = "LambdaX", size = NULL, use_names = TRUE,
  summary = FALSE, iter = NULL, from_scratch = FALSE, ...)

Arguments

object
An object of class pibblefit

newdata
An optional matrix for which to evaluate predictions. If NULL (default), the original data of the model is used.

response
Options = "LambdaX":Mean of regression, "Eta", "Y": counts

size
the number of counts per sample if response="Y" (as vector or matrix), default if newdata=NULL and response="Y" is to use colsums of m$Y. Otherwise uses median colsums of m$Y as default. If passed as a matrix should have dimensions ncol(newdata) x iter.

use_names
if TRUE apply names to output

summary
if TRUE, posterior summary of predictions are returned rather than samples

iter
number of iterations to return if NULL uses object$iter

from_scratch
should predictions of Y come from fitted Eta or from predictions of Eta from posterior of Lambda? (default: false)

...
other arguments passed to summarise_posterior

Value

(if summary==FALSE) array D x N x iter; (if summary==TRUE) tibble with calculated posterior summaries

Details

currently only implemented for pibblefit objects in coord_system "default" "alr", or "ilr".

Examples

sim <- pibble_sim()
fit <- pibble(sim$Y, sim$X)
predict(fit)[,,1:2] # just show 2 samples

#> , , 1
#> 
#>                     s1         s2         s3        s4          s5         s6
#> log(c1/c10) -1.5807781 -0.9490474  0.1406661  1.139274 -0.51996970 -0.1803407
#> log(c2/c10) -0.7532825 -0.2638544  0.5803921  1.354055  0.06856997  0.3316947
#> log(c3/c10)  3.3966648  2.1920883  0.1142357 -1.789898  1.37392814  0.7263276
#> log(c4/c10)  1.2448220  1.0621053  0.7469254  0.458096  0.93800235  0.8397708
#> log(c5/c10)  0.4112773 -0.5859993 -2.3062666 -3.882711 -1.26335925 -1.7995118
#> log(c6/c10) -0.6614721  0.2162113  1.7301846  3.117582  0.81234243  1.2841997
#> log(c7/c10)  0.1359209  0.6516331  1.5412189  2.356431  1.00190996  1.2791655
#> log(c8/c10) -0.7906220 -1.4506892 -2.5892822 -3.632683 -1.89901332 -2.2538765
#> log(c9/c10) -0.3900432  0.1074507  0.9656105  1.752024  0.44535346  0.7128146
#>                     s7         s8        s9       s10        s11        s12
#> log(c1/c10)  0.7983584 -1.9904552 -4.229246 -4.926877 -1.5213114  0.9340673
#> log(c2/c10)  1.0899336 -1.0706764 -2.805161 -3.345645 -0.7072111  1.1950729
#> log(c3/c10) -1.1398438  4.1778321  8.446732  9.776965  3.2832743 -1.3986119
#> log(c4/c10)  0.5566998  1.3633137  2.010844  2.212621  1.2276224  0.5174484
#> log(c5/c10) -3.3445270  1.0580107  4.592260  5.693569  0.3174006 -3.5587628
#> log(c6/c10)  2.6439371 -1.2306494 -4.341072 -5.310312 -0.5788531  2.8324818
#> log(c7/c10)  2.0781247 -0.1985183 -2.026151 -2.595661  0.1844664  2.1889105
#> log(c8/c10) -3.2764753 -0.3625687  1.976644  2.705567 -0.8527562 -3.4182715
#> log(c9/c10)  1.4835493 -0.7126679 -2.475737 -3.025128 -0.3432126  1.5904214
#>                    s13        s14        s15        s16         s17        s18
#> log(c1/c10) -2.4866420 -2.8645446 -0.1517397  1.0106013 -0.49192910 -2.0557601
#> log(c2/c10) -1.4550929 -1.7478698  0.3538531  1.2543670  0.09029418 -1.1212707
#> log(c3/c10)  5.1239550  5.8445351  0.6717916 -1.5445461  1.32046068  4.3023546
#> log(c4/c10)  1.5068268  1.6161283  0.8314985  0.4953123  0.92989211  1.3822020
#> log(c5/c10)  1.8413118  2.4378847 -1.8446626 -3.6795827 -1.30762530  1.1611036
#> log(c6/c10) -1.9200172 -2.4450492  1.3239360  2.9388129  0.85130011 -1.3213794
#> log(c7/c10) -0.6035795 -0.9120796  1.3025139  2.2513889  1.02480084 -0.2518298
#> log(c8/c10)  0.1558746  0.5507281 -2.2837604 -3.4982385 -1.92831168 -0.2943347
#> log(c9/c10) -1.1034196 -1.4010215  0.7353381  1.6506927  0.46743570 -0.7640960
#>                   s19        s20        s21       s22        s23        s24
#> log(c1/c10) -3.574403  0.5408134 -0.2895110 -4.814855  0.4001066 -2.1376157
#> log(c2/c10) -2.297827  0.8904028  0.2471159 -3.258856  0.7813913 -1.1846877
#> log(c3/c10)  7.198084 -0.6487602  0.9344923  9.563363 -0.3804621  4.4584359
#> log(c4/c10)  1.821442  0.6311900  0.8713463  2.180220  0.6718869  1.4058773
#> log(c5/c10)  3.558497 -2.9379558 -1.6271710  5.516726 -2.7158301  1.2903244
#> log(c6/c10) -3.431278  2.2861218  1.1325259 -5.154677  2.0906333 -1.4351040
#> log(c7/c10) -1.491571  1.8678784  1.1900445 -2.504212  1.7530126 -0.3186525
#> log(c8/c10)  1.292427 -3.0073781 -2.1398093  2.588520 -2.8603598 -0.2088074
#> log(c9/c10) -1.960042  1.2807302  0.6268419 -2.936909  1.1699222 -0.8285581
#>                   s25        s26        s27        s28         s29        s30
#> log(c1/c10) -5.093333  1.6879659 -3.0475694 -2.2173792 -1.17347896 -1.4731881
#> log(c2/c10) -3.474605  1.7791496 -1.8896667 -1.2464839 -0.43773085 -0.6699280
#> log(c3/c10) 10.094362 -2.8361366  6.1935245  4.6105281  2.62003167  3.1915134
#> log(c4/c10)  2.260765  0.2993968  1.6690649  1.4289474  1.12701811  1.2137036
#> log(c5/c10)  5.956343 -4.7488987  2.7268153  1.4162424 -0.23170220  0.2414311
#> log(c6/c10) -5.541575  3.8798968 -2.6993313 -1.5459220 -0.09559854 -0.5119939
#> log(c7/c10) -2.731547  2.8043543 -1.0614916 -0.3837673  0.46841883  0.2237518
#> log(c8/c10)  2.879489 -4.2059865  0.7419625 -0.1254660 -1.21619072 -0.9030381
#> log(c9/c10) -3.156214  2.1841237 -1.5451553 -0.8913727 -0.06929128 -0.3053151
#> 
#> , , 2
#> 
#>                      s1         s2         s3        s4         s5          s6
#> log(c1/c10) -1.82116230 -0.9670895  0.5061563  1.856231 -0.3869950  0.07216888
#> log(c2/c10) -0.60171865 -0.1281790  0.6886604  1.437208  0.1934538  0.44803661
#> log(c3/c10)  2.55385432  1.5778980 -0.1055928 -1.648335  0.9150189  0.39032839
#> log(c4/c10)  1.35207547  1.2902577  1.1836242  1.085906  1.2482705  1.21503619
#> log(c5/c10) -0.06143317 -0.7826831 -2.0268141 -3.166930 -1.2725631 -1.66031915
#> log(c6/c10) -0.69417563  0.1477212  1.5999638  2.930792  0.7195457  1.17216350
#> log(c7/c10)  1.20922266  1.3060222  1.4729979  1.626014  1.3717693  1.42381035
#> log(c8/c10) -0.58234096 -1.2797561 -2.4827730 -3.585212 -1.7534473 -2.12838938
#> log(c9/c10) -1.48082961 -0.8932384  0.1203360  1.049170 -0.4941407 -0.17824180
#>                     s7         s8         s9        s10        s11        s12
#> log(c1/c10)  1.3953278 -2.3750281 -5.4017771 -6.3449439 -1.7407658  1.5788005
#> log(c2/c10)  1.1816605 -0.9088089 -2.5869864 -3.1099241 -0.5571429  1.2833867
#> log(c3/c10) -1.1216568  3.1867616  6.6454539  7.7232187  2.4619845 -1.3313126
#> log(c4/c10)  1.1192660  1.3921643  1.6112403  1.6795067  1.3462564  1.1059862
#> log(c5/c10) -2.7777044  0.4062971  2.9623351  3.7588235 -0.1293266 -2.9326439
#> log(c6/c10)  2.4764590 -1.2401453 -4.2237439 -5.1534645 -0.6149253  2.6573160
#> log(c7/c10)  1.5737755  1.1464482  0.8034004  0.6965031  1.2183347  1.5945700
#> log(c8/c10) -3.2088490 -0.1300676  2.3415025  3.1116696 -0.6479908 -3.3586683
#> log(c9/c10)  0.7320747 -1.8618822 -3.9442471 -4.5931339 -1.4255178  0.8583016
#>                    s13        s14        s15        s16        s17         s18
#> log(c1/c10) -3.0458512 -3.5567593  0.1108362  1.6822712 -0.3490853 -2.46331740
#> log(c2/c10) -1.2807460 -1.5640184  0.4694757  1.3407560  0.2144727 -0.95776080
#> log(c3/c10)  3.9533169  4.5371360  0.3461429 -1.4495494  0.8716992  3.28765053
#> log(c4/c10)  1.4407184  1.4776980  1.2122374  1.0984970  1.2455266  1.39855465
#> log(c5/c10)  0.9727958  1.4042490 -1.6929731 -3.0200232 -1.3045772  0.48085590
#> log(c6/c10) -1.9014049 -2.4050293  1.2102796  2.7593116  0.7569149 -1.32717593
#> log(c7/c10)  1.0704180  1.0125124  1.4281929  1.6062973  1.3760660  1.13644163
#> log(c8/c10)  0.4177104  0.8349055 -2.1599642 -3.4431600 -1.7844034 -0.05797264
#> log(c9/c10) -2.3234000 -2.6748983 -0.1516392  0.9294881 -0.4680593 -1.92262417
#>                    s19        s20         s21        s22        s23         s24
#> log(c1/c10) -4.5164572  1.0471381 -0.07542475 -6.1934950  0.8569085 -2.57398270
#> log(c2/c10) -2.0961218  0.9886072  0.36620350 -3.0259534  0.8831346 -1.01911905
#> log(c3/c10)  5.6337912 -0.7237774  0.55898490  7.5501567 -0.5064004  3.41410872
#> log(c4/c10)  1.5471609  1.1444679  1.22571901  1.6685448  1.1582367  1.40656460
#> log(c5/c10)  2.2146976 -2.4836641 -1.53567884  3.6309274 -2.3230184  0.57431086
#> log(c6/c10) -3.3510454  2.1332332  1.02667403 -5.0041747  1.9457156 -1.43626355
#> log(c7/c10)  0.9037414  1.5343121  1.40708228  0.7136682  1.5127517  1.12389897
#> log(c8/c10)  1.6185717 -2.9245257 -2.00786800  2.9880001 -2.7691888  0.03239397
#> log(c9/c10) -3.3351583  0.4925247 -0.27978434 -4.4889390  0.3616491 -1.99876049
#>                    s25       s26        s27        s28        s29        s30
#> log(c1/c10) -6.5699852  2.598039 -3.8042009 -2.6818196 -1.2705113 -1.6757052
#> log(c2/c10) -3.2346980  1.848502 -1.7012121 -1.0789091 -0.2964109 -0.5210701
#> log(c3/c10)  7.9803754 -2.496005  4.8198897  3.5373349  1.9246207  2.3876392
#> log(c4/c10)  1.6957952  1.032214  1.4956078  1.4143698  1.3122194  1.3415473
#> log(c5/c10)  3.9488670 -3.793373  1.6132092  0.6653773 -0.5264485 -0.1842692
#> log(c6/c10) -5.3752975  3.662024 -2.6489433 -1.5425631 -0.1513750 -0.5507922
#> log(c7/c10)  0.6709972  1.710089  0.9844676  1.1116769  1.2716327  1.2257086
#> log(c8/c10)  3.2954329 -4.190954  1.0369604  0.1204510 -1.0319892 -0.7011177
#> log(c9/c10) -4.7479595  1.559525 -2.8451350 -2.0729509 -1.1019887 -1.3807570
#>