Predict response from new data

Predict response from new data

# S3 method for pibblefit
predict(object, newdata = NULL,
  response = "LambdaX", size = NULL, use_names = TRUE,
  summary = FALSE, iter = NULL, from_scratch = FALSE, ...)

Arguments

object
An object of class pibblefit

newdata
An optional matrix for which to evaluate predictions. If NULL (default), the original data of the model is used.

response
Options = "LambdaX":Mean of regression, "Eta", "Y": counts

size
the number of counts per sample if response="Y" (as vector or matrix), default if newdata=NULL and response="Y" is to use colsums of m$Y. Otherwise uses median colsums of m$Y as default. If passed as a matrix should have dimensions ncol(newdata) x iter.

use_names
if TRUE apply names to output

summary
if TRUE, posterior summary of predictions are returned rather than samples

iter
number of iterations to return if NULL uses object$iter

from_scratch
should predictions of Y come from fitted Eta or from predictions of Eta from posterior of Lambda? (default: false)

...
other arguments passed to summarise_posterior

Value

(if summary==FALSE) array D x N x iter; (if summary==TRUE) tibble with calculated posterior summaries

Details

currently only implemented for pibblefit objects in coord_system "default" "alr", or "ilr".

Examples

sim <- pibble_sim()
fit <- pibble(sim$Y, sim$X)
predict(fit)[,,1:2] # just show 2 samples

#> , , 1
#> 
#>                     s1         s2         s3         s4         s5          s6
#> log(c1/c10) -1.4538055 -0.8890208 -0.9621368  1.3078195 -1.7354638 -1.39997313
#> log(c2/c10) -0.1115782  0.1343052  0.1024736  1.0907167 -0.2342003 -0.08814189
#> log(c3/c10)  1.7876498  1.9676741  1.9443685  2.6679140  1.6978716  1.80480874
#> log(c4/c10)  1.7347540  1.5487931  1.5728672  0.8254620  1.8274927  1.71702912
#> log(c5/c10) -3.5836256 -2.6314879 -2.7547499  1.0720372 -4.0584568 -3.49287270
#> log(c6/c10)  0.7899975 -0.7287345 -0.5321223 -6.6361388  1.5473895  0.64523983
#> log(c7/c10)  0.4011195 -0.4382749 -0.3296085 -3.7032631  0.8197256  0.32111272
#> log(c8/c10)  3.8973370  4.4194020  4.3518165  6.4500756  3.6369831  3.94709758
#> log(c9/c10) -3.6824683 -2.8713937 -2.9763939  0.2834393 -4.0869514 -3.60516089
#>                     s7         s8         s9        s10         s11         s12
#> log(c1/c10) -1.5154154  0.9635034 -0.6493737  0.5139761 -1.16182838 -0.00713831
#> log(c2/c10) -0.1384006  0.9408160  0.2386374  0.7451108  0.01553628  0.51823957
#> log(c3/c10)  1.7680117  2.5581637  2.0440613  2.4148774  1.88071706  2.24877294
#> log(c4/c10)  1.7550396  0.9388314  1.4698869  1.0868427  1.63861764  1.25842467
#> log(c5/c10) -3.6874901  0.4915747 -2.2274809 -0.2662572 -3.09139833 -1.14477361
#> log(c6/c10)  0.9556694 -5.7102569 -1.3731564 -4.5014574  0.00485767 -3.10015687
#> log(c7/c10)  0.4926854 -3.1915335 -0.7944432 -2.5234370 -0.03282279 -1.74894634
#> log(c8/c10)  3.8403872  6.1318032  4.6409224  5.7162777  4.16722932  5.23457983
#> log(c9/c10) -3.7709450 -0.2110254 -2.5272419 -0.8565814 -3.26316656 -1.60494211
#>                    s13        s14        s15        s16        s17         s18
#> log(c1/c10) -0.3587709 -0.7476966  0.8145512 -0.8306206 -1.7338543 -1.38885896
#> log(c2/c10)  0.3651536  0.1958317  0.8759685  0.1597302 -0.2334997 -0.08330325
#> log(c3/c10)  2.1366905  2.0127210  2.5106853  1.9862891  1.6983846  1.80835137
#> log(c4/c10)  1.3742031  1.5022607  0.9878754  1.5295642  1.8269628  1.71336967
#> log(c5/c10) -1.7375705 -2.3932378  0.2404648 -2.5330344 -4.0557435 -3.47413597
#> log(c6/c10) -2.1546007 -1.1087616 -5.3097175 -0.8857752  1.5430616  0.61535332
#> log(c7/c10) -1.2263429 -0.6483137 -2.9701577 -0.5250704  0.8173336  0.30459462
#> log(c8/c10)  4.9095444  4.5500366  5.9941176  4.4733849  3.6384708  3.95737109
#> log(c9/c10) -2.1099138 -2.6684413 -0.4249323 -2.7875265 -4.0846401 -3.58920008
#>                    s19          s20        s21         s22        s23
#> log(c1/c10) -0.7930844 -1.179108206 -2.0931468 -0.04342092 -0.8782683
#> log(c2/c10)  0.1760718  0.008013370 -0.3899205  0.50244366  0.1389863
#> log(c3/c10)  1.9982537  1.875209137  1.5838606  2.23720791  1.9711014
#> log(c4/c10)  1.5172051  1.644307188  1.9452633  1.27037107  1.5452527
#> log(c5/c10) -2.4697543 -3.120529384 -4.6614539 -1.20594034 -2.6133610
#> log(c6/c10) -0.9867118  0.051323915  2.5092157 -3.00259128 -0.7576483
#> log(c7/c10) -0.5808575 -0.007141163  1.3513214 -1.69502240 -0.4542554
#> log(c8/c10)  4.5080819  4.151256518  3.3063548  5.20104160  4.4293412
#> log(c9/c10) -2.7336216 -3.287981731 -4.6006121 -1.65704674 -2.8559523
#>                    s24         s25          s26        s27        s28
#> log(c1/c10) -2.4665873  0.02917486  0.001605888  0.1050571 -0.5047816
#> log(c2/c10) -0.5525007  0.53404879  0.522046428  0.5670847  0.3015867
#> log(c3/c10)  1.4648269  2.26034771  2.251560139  2.2845351  2.0901498
#> log(c4/c10)  2.0682222  1.24646821  1.255545560  1.2214832  1.4222786
#> log(c5/c10) -5.2910154 -1.08355536 -1.130032276 -0.9556299 -1.9837214
#> log(c6/c10)  3.5134143 -3.19780464 -3.123670416 -3.4018555 -1.7619713
#> log(c7/c10)  1.9063361 -1.80291569 -1.761942137 -1.9156934 -1.0093389
#> log(c8/c10)  2.9611610  5.26814631  5.242662628  5.3382889  4.7745778
#> log(c9/c10) -5.1369015 -1.55279358 -1.592384761 -1.4438208 -2.3195964
#>                    s29        s30
#> log(c1/c10) -2.9438091 -1.7131311
#> log(c2/c10) -0.7602629 -0.2244777
#> log(c3/c10)  1.3127131  1.7049901
#> log(c4/c10)  2.2253522  1.8201394
#> log(c5/c10) -6.0955359 -4.0208075
#> log(c6/c10)  4.7966857  1.4873361
#> log(c7/c10)  2.6155928  0.7865344
#> log(c8/c10)  2.5200357  3.6576265
#> log(c9/c10) -5.8222291 -4.0548801
#> 
#> , , 2
#> 
#>                       s1          s2         s3         s4         s5
#> log(c1/c10) -1.410176580 -0.36454788 -0.4999130  3.7026288 -1.9316318
#> log(c2/c10) -0.008451428  0.20062223  0.1735559  1.0138550 -0.1127165
#> log(c3/c10)  2.183056344  2.22156675  2.2165813  2.3713605  2.1638512
#> log(c4/c10)  2.110561096  1.53223056  1.6071002 -0.7172988  2.3989746
#> log(c5/c10) -3.200315133 -2.25694880 -2.3790753  1.4124584 -3.6707721
#> log(c6/c10)  0.956035456 -1.26795865 -0.9800445 -9.9186172  2.0651417
#> log(c7/c10)  0.431783674  0.06467993  0.1122045 -1.3632416  0.6148584
#> log(c8/c10)  4.746173941  5.10931231  5.0623011  6.5218098  4.5650768
#> log(c9/c10) -3.228421417 -2.50823097 -2.6014655  0.2930902 -3.5875805
#>                     s6          s7         s8          s9         s10
#> log(c1/c10) -1.3105126 -1.52423967  3.0651703  0.07912884  2.23292636
#> log(c2/c10)  0.0114764 -0.03125836  0.8863950  0.28933548  0.71998771
#> log(c3/c10)  2.1867270  2.17885541  2.3478829  2.23790732  2.31723146
#> log(c4/c10)  2.0554376  2.17364866 -0.3647246  1.28683581  0.09558419
#> log(c5/c10) -3.1103983 -3.30322286  0.8373433 -1.85666362  0.08649274
#> log(c6/c10)  0.7440558  1.19864130 -8.5627784 -2.21163434 -6.79264189
#> log(c7/c10)  0.3967932  0.47182942 -1.1394400 -0.09108797 -0.84725230
#> log(c8/c10)  4.7807864  4.70656076  6.3004256  5.26339764  6.01139406
#> log(c9/c10) -3.1597766 -3.30698386 -0.1459677 -2.20264282 -0.71918660
#>                     s11        s12        s13         s14        s15
#> log(c1/c10) -0.86961724  1.2681478  0.6171438 -0.10290383  2.7894037
#> log(c2/c10)  0.09963352  0.5270801  0.3969117  0.25293801  0.8312555
#> log(c3/c10)  2.20296509  2.2816988  2.2577224  2.23120307  2.3377265
#> log(c4/c10)  1.81158119  0.6291970  0.9892632  1.38751691 -0.2121999
#> log(c5/c10) -2.71262244 -0.7839305 -1.3712665 -2.02089352  0.5885467
#> log(c6/c10) -0.19370419 -4.7406116 -3.3559622 -1.82446097 -7.9762382
#> log(c7/c10)  0.24200182 -0.5085338 -0.2799765 -0.02717917 -1.0426227
#> log(c8/c10)  4.93390582  5.6763343  5.4502459  5.20017917  6.2046541
#> log(c9/c10) -2.85610410 -1.3836904 -1.8320780 -2.32802021 -0.3359056
#>                     s16        s17         s18          s19         s20
#> log(c1/c10) -0.25642724 -1.9286521 -1.28993614 -0.186933608 -0.90160869
#> log(c2/c10)  0.22224098 -0.1121207  0.01559067  0.236136242  0.09323682
#> log(c3/c10)  2.22554882  2.1639609  2.18748479  2.228108261  2.20178685
#> log(c4/c10)  1.47242972  2.3973266  2.04405688  1.433993243  1.82927545
#> log(c5/c10) -2.15940235 -3.6680838 -3.09183419 -2.096705190 -2.74148513
#> log(c6/c10) -1.49792523  2.0588041  0.70029072 -1.645734313 -0.12566015
#> log(c7/c10)  0.02672048  0.6138123  0.38956915  0.002322363  0.25323351
#> log(c8/c10)  5.14686173  4.5661116  4.78793247  5.170996309  4.92279545
#> log(c9/c10) -2.43376146 -3.5855282 -3.14560424 -2.385896817 -2.87813863
#>                    s21        s22         s23       s24        s25        s26
#> log(c1/c10) -2.5938376  1.2009751 -0.34464107 -3.285216  1.3353772  1.2843366
#> log(c2/c10) -0.2451247  0.5136489  0.20460260 -0.383366  0.5405226  0.5303170
#> log(c3/c10)  2.1394622  2.2792248  2.22229991  2.113999  2.2841748  2.2822950
#> log(c4/c10)  2.7652364  0.6663498  1.52122023  3.147633  0.5920129  0.6202431
#> log(c5/c10) -4.2682142 -0.8445338 -2.23898888 -4.891976 -0.7232763 -0.7693250
#> log(c6/c10)  3.4736166 -4.5977390 -1.31029932  4.944140 -4.8836046 -4.7750443
#> log(c7/c10)  0.8473484 -0.4849505  0.05769096  1.090080 -0.5321369 -0.5142174
#> log(c8/c10)  4.3350981  5.6530058  5.11622579  4.094988  5.6996825  5.6819565
#> log(c9/c10) -4.0436834 -1.4299565 -2.49451989 -4.519879 -1.3373854 -1.3725402
#>                    s27        s28        s29        s30
#> log(c1/c10)  1.4758637  0.3468232 -4.1687331 -1.8902858
#> log(c2/c10)  0.5686129  0.3428610 -0.5600253 -0.1044493
#> log(c3/c10)  2.2893489  2.2477665  2.0814590  2.1653740
#> log(c4/c10)  0.5143107  1.1387758  3.6363010  2.3761064
#> log(c5/c10) -0.5965293 -1.6151497 -5.6890851 -3.6334697
#> log(c6/c10) -5.1824117 -2.7810054  6.8233317  1.9772010
#> log(c7/c10) -0.5814595 -0.1850712  1.4002693  0.6003424
#> log(c8/c10)  5.7484723  5.3563657  3.7881496  4.5794359
#> log(c9/c10) -1.2406234 -2.0182648 -5.1284133 -3.5591029
#>