Predict response from new data

Predict response from new data

# S3 method for pibblefit
predict(object, newdata = NULL,
  response = "LambdaX", size = NULL, use_names = TRUE,
  summary = FALSE, iter = NULL, from_scratch = FALSE, ...)

Arguments

object
An object of class pibblefit

newdata
An optional matrix for which to evaluate predictions. If NULL (default), the original data of the model is used.

response
Options = "LambdaX":Mean of regression, "Eta", "Y": counts

size
the number of counts per sample if response="Y" (as vector or matrix), default if newdata=NULL and response="Y" is to use colsums of m$Y. Otherwise uses median colsums of m$Y as default. If passed as a matrix should have dimensions ncol(newdata) x iter.

use_names
if TRUE apply names to output

summary
if TRUE, posterior summary of predictions are returned rather than samples

iter
number of iterations to return if NULL uses object$iter

from_scratch
should predictions of Y come from fitted Eta or from predictions of Eta from posterior of Lambda? (default: false)

...
other arguments passed to summarise_posterior

Value

(if summary==FALSE) array D x N x iter; (if summary==TRUE) tibble with calculated posterior summaries

Details

currently only implemented for pibblefit objects in coord_system "default" "alr", or "ilr".

Examples

sim <- pibble_sim()
fit <- pibble(sim$Y, sim$X)
predict(fit)[,,1:2] # just show 2 samples

#> , , 1
#> 
#>                     s1         s2         s3         s4          s5         s6
#> log(c1/c10) -1.5751161 -0.9473990  0.1353911  1.1276545 -0.52104746 -0.1835762
#> log(c2/c10) -0.7520634 -0.2626917  0.5814573  1.3550313  0.06969429  0.3327887
#> log(c3/c10)  3.3983386  2.1936028  0.1154757 -1.7889096  1.37533458  0.7276485
#> log(c4/c10)  1.2450933  1.0620480  0.7463013  0.4569527  0.93772189  0.8393137
#> log(c5/c10)  0.4108138 -0.5869480 -2.3080525 -3.8852639 -1.26463761 -1.8010511
#> log(c6/c10) -0.6598954  0.2166836  1.7287517  3.1144034  0.81206454  1.2833281
#> log(c7/c10)  0.1380288  0.6530569  1.5414622  2.3555925  1.00286891  1.2797566
#> log(c8/c10) -0.7894097 -1.4498732 -2.5891498 -3.6331772 -1.89846650 -2.2535428
#> log(c9/c10) -0.3811750  0.1121817  0.9632047  1.7430779  0.44727432  0.7125111
#>                     s7         s8        s9       s10        s11        s12
#> log(c1/c10)  0.7889047 -1.9821904 -4.206757 -4.899956 -1.5160272  0.9237514
#> log(c2/c10)  1.0909401 -1.0694207 -2.803705 -3.344126 -0.7059973  1.1960674
#> log(c3/c10) -1.1387696  4.1796091  8.449073  9.779483  3.2849331 -1.3975719
#> log(c4/c10)  0.5557337  1.3637980  2.012492  2.214632  1.2278626  0.5164118
#> log(c5/c10) -3.3468181  1.0578619  4.593831  5.695676  0.3168914 -3.5611582
#> log(c6/c10)  2.6413544 -1.2283564 -4.334865 -5.302886 -0.5773803  2.8296618
#> log(c7/c10)  2.0776557 -0.1959666 -2.021175 -2.589929  0.1865100  2.1882945
#> log(c8/c10) -3.2767556 -0.3610993  1.979518  2.708879 -0.8515811 -3.4186370
#> log(c9/c10)  1.4768362 -0.7011165 -2.449523 -2.994345 -0.3347338  1.5828195
#>                    s13        s14        s15        s16        s17        s18
#> log(c1/c10) -2.4752246 -2.8507262 -0.1551569  0.9997992 -0.4931850 -2.0470803
#> log(c2/c10) -1.4537929 -1.7465360  0.3549445  1.2553546  0.0914160 -1.1200092
#> log(c3/c10)  5.1258570  5.8465324  0.6731052 -1.5435254  1.3218601  4.3041480
#> log(c4/c10)  1.5075691  1.6170672  0.8310265  0.4942359  0.9295971  1.3827202
#> log(c5/c10)  1.8415442  2.4384074 -1.8462238 -3.6820368 -1.3089252  1.1610050
#> log(c6/c10) -1.9168568 -2.4412281  1.3230143  2.9358591  0.8509732 -1.3189723
#> log(c7/c10) -0.6004903 -0.9085811  1.3030740  2.2506900  1.0257294 -0.2492073
#> log(c8/c10)  0.1576553  0.5527459 -2.2834446 -3.4986520 -1.9277824 -0.2928243
#> log(c9/c10) -1.0886187 -1.3837456  0.7348474  1.6425895  0.4691729 -0.7521170
#>                   s19        s20        s21       s22        s23        s24
#> log(c1/c10) -3.556075  0.5329960 -0.2920530 -4.788645  0.3931831 -2.1284158
#> log(c2/c10) -2.296429  0.8914323  0.2482196 -3.257348  0.7824334 -1.1834189
#> log(c3/c10)  7.200261 -0.6476211  0.9358407  9.565852 -0.3792875  4.4602500
#> log(c4/c10)  1.822750  0.6303579  0.8709460  2.182174  0.6711280  1.4064381
#> log(c5/c10)  3.559565 -2.9400491 -1.6286264  5.518747 -2.7178152  1.2902887
#> log(c6/c10) -3.426216  2.2839893  1.1318451 -5.147446  2.0887469 -1.4325538
#> log(c7/c10) -1.487304  1.8676883  1.1907538 -2.498601  1.7529749 -0.3159414
#> log(c8/c10)  1.294891 -3.0074968 -2.1394071  2.591762 -2.8603903 -0.2072456
#> log(c9/c10) -1.938117  1.2757038  0.6272534 -2.906861  1.1658174 -0.8160430
#>                   s25        s26        s27        s28         s29        s30
#> log(c1/c10) -5.065354  1.6728601 -3.0325882 -2.2076726 -1.17040469 -1.4682096
#> log(c2/c10) -3.473072  1.7800766 -1.8883165 -1.2452079 -0.43654815 -0.6687185
#> log(c3/c10) 10.096921 -2.8352866  6.1955679  4.6123623  2.62160280  3.1931601
#> log(c4/c10)  2.262863  0.2979681  1.6700990  1.4295497  1.12707751  1.2139188
#> log(c5/c10)  5.958578 -4.7518731  2.7274786  1.4162680 -0.23247855  0.2408850
#> log(c6/c10) -5.533857  3.8757588 -2.6951902 -1.5432324 -0.09473391 -0.5106052
#> log(c7/c10) -2.725634  2.8029217 -1.0577948 -0.3809698  0.47008563  0.2257433
#> log(c8/c10)  2.882906 -4.2068249  0.7440951 -0.1238542 -1.21523389 -0.9018932
#> log(c9/c10) -3.124341  2.1715844 -1.5266807 -0.8783352 -0.06309048 -0.2971514
#> 
#> , , 2
#> 
#>                      s1         s2         s3        s4         s5          s6
#> log(c1/c10) -1.81412262 -0.9640974  0.5021666  1.845843 -0.3867519  0.07023589
#> log(c2/c10) -0.60080968 -0.1276646  0.6884942  1.436418  0.1937002  0.44807091
#> log(c3/c10)  2.55412938  1.5782561 -0.1050912 -1.647702  0.9154335  0.39078767
#> log(c4/c10)  1.35163634  1.2895210  1.1823743  1.084186  1.2473317  1.21393750
#> log(c5/c10) -0.06163158 -0.7831952 -2.0278672 -3.168478 -1.2732882 -1.66121285
#> log(c6/c10) -0.69204904  0.1486164  1.5987350  2.927616  0.7196045  1.17156037
#> log(c7/c10)  1.20913798  1.3054657  1.4716276  1.623898  1.3708924  1.42267977
#> log(c8/c10) -0.58266686 -1.2804382 -2.4840694 -3.587071 -1.7543713 -2.12950486
#> log(c9/c10) -1.46740618 -0.8860780  0.1166929  1.035627 -0.4912342 -0.17870244
#>                     s7         s8         s9        s10        s11        s12
#> log(c1/c10)  1.3871243 -2.3653636 -5.3777686 -6.3164656 -1.7341071  1.5697274
#> log(c2/c10)  1.1810836 -0.9076440 -2.5844232 -3.1069252 -0.5562711  1.2827250
#> log(c3/c10) -1.1210688  3.1869827  6.6453805  7.7230536  2.4622674 -1.3307067
#> log(c4/c10)  1.1177063  1.3919181  1.6120485  1.6806435  1.3457892  1.1043627
#> log(c5/c10) -2.7790840  0.4063021  2.9634516  3.7602863 -0.1295546 -2.9340909
#> log(c6/c10)  2.4739482 -1.2372202 -4.2164549 -5.1448157 -0.6129146  2.6545406
#> log(c7/c10)  1.5719139  1.1466695  0.8052937  0.6989175  1.2182056  1.5926071
#> log(c8/c10) -3.2105162 -0.1301625  2.3426698  3.1132303 -0.6483502 -3.3604121
#> log(c9/c10)  0.7219112 -1.8443972 -3.9045666 -4.5465370 -1.4126840  0.8467926
#>                    s13        s14        s15        s16        s17         s18
#> log(c1/c10) -3.0330076 -3.5414944  0.1087200  1.6727078 -0.3490219 -2.45323450
#> log(c2/c10) -1.2792712 -1.5623076  0.4694921  1.3400464  0.2147016 -0.95655517
#> log(c3/c10)  3.9534728  4.5372422  0.3466059 -1.4489335  0.8721175  3.28786310
#> log(c4/c10)  1.4407059  1.4778634  1.2111253  1.0968374  1.2445746  1.39833922
#> log(c5/c10)  0.9730471  1.4046879 -1.6938810 -3.0215081 -1.3053162  0.48089331
#> log(c6/c10) -1.8975126 -2.4004004  1.2096207  2.7563871  0.7569191 -1.32412351
#> log(c7/c10)  1.0710099  1.0133864  1.4270409  1.6042772  1.3751681  1.13671169
#> log(c8/c10)  0.4178952  0.8353034 -2.1610958 -3.4449469 -1.7853432 -0.05803074
#> log(c9/c10) -2.3009958 -2.6487475 -0.1523834  0.9172204 -0.4654308 -1.90449173
#>                    s19        s20         s21        s22        s23         s24
#> log(c1/c10) -4.4966443  1.0405847 -0.07665827 -6.1657344  0.8512566 -2.56337534
#> log(c2/c10) -2.0939677  0.9881911  0.36630599 -3.0230245  0.8828064 -1.01786229
#> log(c3/c10)  5.6338040 -0.7232233  0.55942982  7.5500064 -0.5058648  3.41431052
#> log(c4/c10)  1.5476607  1.1430296  1.22467174  1.6696289  1.1568647  1.40638773
#> log(c5/c10)  2.2154890 -2.4849158 -1.53651834  3.6323347 -2.3242003  0.57438891
#> log(c6/c10) -3.3450329  2.1312244  1.02628369 -4.9957443  1.9439811 -1.43305157
#> log(c7/c10)  0.9051457  1.5326429  1.40603323  0.7159988  1.5111876  1.12423016
#> log(c8/c10)  1.6193698 -2.9260478 -2.00892193  2.9894976 -2.7706315  0.03238201
#> log(c9/c10) -3.3019699  0.4849144 -0.27916265 -4.4434526  0.3554339 -1.97981652
#>                    s25       s26        s27        s28        s29        s30
#> log(c1/c10) -6.5404404  2.584136 -3.7877634 -2.6707012 -1.2660812 -1.6693548
#> log(c2/c10) -3.2315952  1.847370 -1.6993870 -1.0776025 -0.2957563 -0.5202283
#> log(c3/c10)  7.9801884 -2.495300  4.8199718  3.5375262  1.9249493  2.3879284
#> log(c4/c10)  1.6970104  1.030235  1.4958595  1.4142305  1.3115884  1.3410575
#> log(c5/c10)  3.9504125 -3.795195  1.6137390  0.6654949 -0.5268492 -0.1845211
#> log(c6/c10) -5.3663243  3.657779 -2.6439576 -1.5391956 -0.1500423 -0.5488753
#> log(c7/c10)  0.6735359  1.707563  0.9854784  1.1120676  1.2712439  1.2255435
#> log(c8/c10)  3.2970874 -4.193123  1.0374614  0.1204840 -1.0325447 -0.7015043
#> log(c9/c10) -4.6997123  1.540542 -2.8171696 -2.0532161 -1.0926033 -1.3684002
#>