Predict response from new data

Predict response from new data

# S3 method for pibblefit
predict(object, newdata = NULL,
  response = "LambdaX", size = NULL, use_names = TRUE,
  summary = FALSE, iter = NULL, from_scratch = FALSE, ...)

Arguments

object
An object of class pibblefit

newdata
An optional matrix for which to evaluate predictions. If NULL (default), the original data of the model is used.

response
Options = "LambdaX":Mean of regression, "Eta", "Y": counts

size
the number of counts per sample if response="Y" (as vector or matrix), default if newdata=NULL and response="Y" is to use colsums of m$Y. Otherwise uses median colsums of m$Y as default. If passed as a matrix should have dimensions ncol(newdata) x iter.

use_names
if TRUE apply names to output

summary
if TRUE, posterior summary of predictions are returned rather than samples

iter
number of iterations to return if NULL uses object$iter

from_scratch
should predictions of Y come from fitted Eta or from predictions of Eta from posterior of Lambda? (default: false)

...
other arguments passed to summarise_posterior

Value

(if summary==FALSE) array D x N x iter; (if summary==TRUE) tibble with calculated posterior summaries

Details

currently only implemented for pibblefit objects in coord_system "default" "alr", or "ilr".

Examples

sim <- pibble_sim()
fit <- pibble(sim$Y, sim$X)
predict(fit)[,,1:2] # just show 2 samples

#> , , 1
#> 
#>                      s1          s2         s3         s4          s5
#> log(c1/c10)  1.25832832 -0.06294866  0.4012999  1.5556545  2.73583640
#> log(c2/c10) -2.92115004 -0.46750385 -1.3296256 -3.4732925 -5.66492151
#> log(c3/c10) -1.91932333 -1.88860103 -1.8993957 -1.9262367 -1.95367830
#> log(c4/c10)  0.07690713  3.05433772  2.0081772 -0.5931022 -3.25258234
#> log(c5/c10) -0.63574481 -2.95825150 -2.1422074 -0.1131126  1.96138089
#> log(c6/c10)  0.39928659  0.84524937  0.6885543  0.2989319 -0.09940795
#> log(c7/c10)  0.24575595  0.20542684  0.2195970  0.2548312  0.29085366
#> log(c8/c10)  1.25046307  2.71105479  2.1978561  0.9217870 -0.38283267
#> log(c9/c10)  0.70048151 -0.87831231 -0.3235817  1.0557565  2.46595589
#>                     s6         s7         s8          s9        s10        s11
#> log(c1/c10)  3.3558021  0.6814275  2.1401754  2.57793548  0.1166004  3.2138755
#> log(c2/c10) -6.8162141 -1.8498299 -4.5587633 -5.37169532 -0.8009312 -6.5526528
#> log(c3/c10) -1.9680937 -1.9059092 -1.9398280 -1.95000680 -1.8927759 -1.9647937
#> log(c4/c10) -4.6496435  1.3769239 -1.9102904 -2.89676062  2.6497330 -4.3298193
#> log(c5/c10)  3.0511406 -1.6498063  0.9143432  1.68382670 -2.6426447  2.8016658
#> log(c6/c10) -0.3086613  0.5940045  0.1016420 -0.04611259  0.7846472 -0.2607577
#> log(c7/c10)  0.3097768  0.2281473  0.2726724  0.28603408  0.2109072  0.3054448
#> log(c8/c10) -1.0681672  1.8881920  0.2756345 -0.20828275  2.5125741 -0.9112760
#> log(c9/c10)  3.2067528  0.0111427  1.7542005  2.27728007 -0.6637691  3.0371648
#>                    s12         s13        s14         s15        s16        s17
#> log(c1/c10)  1.9871142  1.32991652  3.4953189  2.55984137  0.2509735  0.8667498
#> log(c2/c10) -4.2745246 -3.05409121 -7.0753006 -5.33809409 -1.0504656 -2.1939784
#> log(c3/c10) -1.9362690 -1.92098789 -1.9713378 -1.94958607 -1.8959003 -1.9102184
#> log(c4/c10) -1.5653748 -0.08441324 -4.9640377 -2.85598647  2.3469301  0.9593094
#> log(c5/c10)  0.6452962 -0.50990892  3.2963797  1.65202135 -2.4064472 -1.3240514
#> log(c6/c10)  0.1533038  0.37512385 -0.3557516 -0.04000539  0.7392931  0.5314538
#> log(c7/c10)  0.2680005  0.24794102  0.3140352  0.28548180  0.2150086  0.2338039
#> log(c8/c10)  0.4448344  1.17132665 -1.2223946 -0.18828081  2.3640328  1.6833294
#> log(c9/c10)  1.5713077  0.78602225  3.3734615  2.25565943 -0.5032067  0.2325844
#>                     s18        s19        s20        s21        s22        s23
#> log(c1/c10)  1.59921012  0.3977562  1.4410734  0.1800367  0.3379689  0.2422733
#> log(c2/c10) -3.55417652 -1.3230450 -3.2605125 -0.9187341 -1.2120184 -1.0343091
#> log(c3/c10) -1.92724950 -1.8993133 -1.9235725 -1.8942509 -1.8979232 -1.8956980
#> log(c4/c10) -0.69125270  2.0161626 -0.3348996  2.5067825  2.1508902  2.3665355
#> log(c5/c10) -0.03655156 -2.1484363 -0.3145202 -2.5311380 -2.2535288 -2.4217402
#> log(c6/c10)  0.28423078  0.6897504  0.3376057  0.7632360  0.7099300  0.7422296
#> log(c7/c10)  0.25616062  0.2194889  0.2513338  0.2128434  0.2176640  0.2147431
#> log(c8/c10)  0.87363888  2.2017734  1.0484495  2.4424492  2.2678646  2.3736504
#> log(c9/c10)  1.10780117 -0.3278160  0.9188436 -0.5879691 -0.3992559 -0.5136026
#>                     s24        s25        s26         s27        s28        s29
#> log(c1/c10)  2.33989466  1.9683272  1.0661315  2.32876713  0.3724098  1.7237203
#> log(c2/c10) -4.92964726 -4.2396366 -2.5642354 -4.90898313 -1.2759760 -3.7853953
#> log(c3/c10) -1.94447188 -1.9358322 -1.9148544 -1.94421314 -1.8987240 -1.9301446
#> log(c4/c10) -2.36034773 -1.5230392  0.5100130 -2.33527239  2.0732795 -0.9718302
#> log(c5/c10)  1.26540468  0.6122729 -0.9735835  1.24584499 -2.1929896  0.1823093
#> log(c6/c10)  0.03423192  0.1596449  0.4641577  0.03798773  0.6983054  0.2422056
#> log(c7/c10)  0.27876840  0.2674271  0.2398896  0.27842876  0.2187152  0.2599610
#> log(c8/c10)  0.05485696  0.4656023  1.4629251  0.06715777  2.2297923  0.7360004
#> log(c9/c10)  1.99284511  1.5488591  0.4708255  1.97954882 -0.3581025  1.2565784
#>                    s30
#> log(c1/c10)  1.4256442
#> log(c2/c10) -3.2318600
#> log(c3/c10) -1.9232138
#> log(c4/c10) -0.3001307
#> log(c5/c10) -0.3416413
#> log(c6/c10)  0.3428135
#> log(c7/c10)  0.2508629
#> log(c8/c10)  1.0655055
#> log(c9/c10)  0.9004073
#> 
#> , , 2
#> 
#>                     s1         s2         s3         s4         s5         s6
#> log(c1/c10)  1.4162138  0.2403800  0.6535251  1.6808110  2.7310812  3.2828025
#> log(c2/c10) -2.8446286 -1.3447964 -1.8717827 -3.1821350 -4.5218049 -5.2255517
#> log(c3/c10) -2.0208209 -0.6023850 -1.1007717 -2.3400107 -3.6069763 -4.2725306
#> log(c4/c10) -0.6341545  2.7362831  1.5520342 -1.3926020 -4.4031210 -5.9845877
#> log(c5/c10) -0.9147972 -2.0299849 -1.6381486 -0.6638472  0.3322531  0.8555182
#> log(c6/c10)  0.6315266  2.1210940  1.5977143  0.2963301 -1.0341710 -1.7331014
#> log(c7/c10)  0.3598887  0.1071546  0.1959562  0.4167613  0.6425067  0.7610939
#> log(c8/c10)  1.1664654  3.1256196  2.4372443  0.7255982 -1.0243441 -1.9436126
#> log(c9/c10)  0.3103315 -0.6572391 -0.3172701  0.5280633  1.3923101  1.8463108
#>                     s7         s8         s9        s10        s11         s12
#> log(c1/c10)  0.9028169  2.2009892  2.5905617  0.4001647  3.1564989  2.06477660
#> log(c2/c10) -2.1897663 -3.8456471 -4.3425655 -1.5486094 -5.0644455 -3.67190149
#> log(c3/c10) -1.4014982 -2.9675138 -3.4374643 -0.7951371 -4.1201676 -2.80319743
#> log(c4/c10)  0.8374585 -2.8836529 -4.0003325  2.2782724 -5.6225482 -2.49320998
#> log(c5/c10) -1.4017146 -0.1704983  0.1989811 -1.8784414  0.7357290 -0.29968543
#> log(c6/c10)  1.2819070 -0.3626409 -0.8561583  1.9186758 -1.5730978 -0.19008440
#> log(c7/c10)  0.2495390  0.5285686  0.6123034  0.1414988  0.7339461  0.49929101
#> log(c8/c10)  2.0218786 -0.1411138 -0.7902128  2.8593890 -1.7331678  0.08584121
#> log(c9/c10) -0.1121328  0.9561078  1.2766793 -0.5257554  1.7423780  0.84402113
#>                    s13        s14        s15        s16         s17        s18
#> log(c1/c10)  1.4799217  3.4069617  2.5744593  0.5197463  1.06773936  1.7195721
#> log(c2/c10) -2.9258912 -5.3839227 -4.3220261 -1.7011415 -2.40013283 -3.2315767
#> log(c3/c10) -2.0976733 -4.4223067 -3.4180396 -0.9393912 -1.60044806 -2.3867692
#> log(c4/c10) -0.8167684 -6.3404805 -3.9541764  1.9355009  0.36472079 -1.5037078
#> log(c5/c10) -0.8543752  0.9732736  0.1837093 -1.7650275 -1.24529835 -0.6270852
#> log(c6/c10)  0.5508202 -1.8903885 -0.8357596  1.7671877  1.07298024  0.2472268
#> log(c7/c10)  0.3735821  0.7877807  0.6088424  0.1672017  0.28498751  0.4250927
#> log(c8/c10)  1.0603163 -2.1504845 -0.7633834  2.6601442  1.74708757  0.6610150
#> log(c9/c10)  0.3627555  1.9484789  1.2634290 -0.4273540  0.02357872  0.5599591
#>                    s19        s20        s21        s22        s23          s24
#> log(c1/c10)  0.6503715  1.5788427  0.4566181  0.5971655  0.5120038  2.378723831
#> log(c2/c10) -1.8677602 -3.0520696 -1.6206184 -1.7998933 -1.6912655 -4.072356186
#> log(c3/c10) -1.0969675 -2.2170040 -0.8632381 -1.0327838 -0.9300512 -3.181919320
#> log(c4/c10)  1.5610737 -1.1003178  2.1164531  1.7135847  1.9576942 -3.393115620
#> log(c5/c10) -1.6411395 -0.7605562 -1.8248998 -1.6916014 -1.7723707 -0.001930706
#> log(c6/c10)  1.6017093  0.4255053  1.8471597  1.6691117  1.7769960 -0.587798372
#> log(c7/c10)  0.1952783  0.3948442  0.1536329  0.1838422  0.1655375  0.566770923
#> log(c8/c10)  2.4424987  0.8954959  2.7653274  2.5311497  2.6730447 -0.437252280
#> log(c9/c10) -0.3198651  0.4441557 -0.4793010 -0.3636473 -0.4337252  1.102362149
#>                    s25        s26        s27        s28        s29        s30
#> log(c1/c10)  2.0480576  1.2451735  2.3688212  0.6278152  1.8303766  1.5651119
#> log(c2/c10) -3.6505756 -2.6264586 -4.0597249 -1.8389885 -3.3729130 -3.0345553
#> log(c3/c10) -2.7830289 -1.8144911 -3.1699735 -1.0697572 -2.5204352 -2.2004403
#> log(c4/c10) -2.4452863 -0.1438806 -3.3647305  1.6257297 -1.8213202 -1.0609596
#> log(c5/c10) -0.3155421 -1.0770157 -0.0113226 -1.6625325 -0.5219958 -0.7735788
#> log(c6/c10) -0.1689045  0.8482035 -0.5752535  1.6302841  0.1068578  0.4428997
#> log(c7/c10)  0.4956974  0.3231253  0.5646424  0.1904301  0.4489090  0.3918929
#> log(c8/c10)  0.1136981  1.4514498 -0.4207527  2.4800817  0.4763946  0.9183739
#> log(c9/c10)  0.8302634  0.1695858  1.0942135 -0.3384263  0.6511379  0.4328569
#>