Provide random initialization for pibble model

Randomly initializes based on ALR transform of counts plus random pseudocounts uniformily distributed between 0 and 1.

random_pibble_init(Y)

Arguments

Y
matrix (D x N) of counts

Value

(D-1) x N matrix

Details

Notation: N is number of samples and D is number of multinomial categories

Examples

Y <- matrix(sample(1:100, 100), 10, 10)
random_pibble_init(Y)

#>            [,1]       [,2]         [,3]        [,4]        [,5]        [,6]
#>  [1,] 1.7417322  0.6580854 -2.820128844  0.04304189  1.76351914 -3.19121632
#>  [2,] 1.3648604  0.1433480 -0.008378035  0.08881165  2.18360804 -1.92381883
#>  [3,] 0.5981767  0.9055992  0.472272399 -0.29194118  0.31376327 -0.85446630
#>  [4,] 1.7966509 -1.7492302  0.438460344 -2.05646334  1.14233930 -0.50457435
#>  [5,] 1.7809859  0.1217808 -1.173436835 -0.61097962  1.74029411 -0.17722418
#>  [6,] 0.8236542  0.8252155  0.177149356 -0.44374554 -0.08763283 -0.27913246
#>  [7,] 0.9846733  0.1663179  0.401647833 -1.49134910  2.09283106 -0.15485619
#>  [8,] 1.2770474 -1.5827391 -0.354031287 -0.27639224  0.66301706  0.03838513
#>  [9,] 0.7038176  0.8361307 -1.018499110 -0.50492226 -2.30923600 -0.07442895
#>             [,7]      [,8]       [,9]       [,10]
#>  [1,]  0.4911615 0.2444067  1.9557669  0.04990157
#>  [2,]  0.6144057 0.3241776  1.1928121  0.20557665
#>  [3,]  0.9440319 0.3769992  0.6668615 -1.55667815
#>  [4,]  0.7776984 0.3849643  0.8897731 -0.59797039
#>  [5,]  0.8448297 1.0776052 -0.8709291 -0.42995060
#>  [6,]  0.3873214 0.7317451  1.3552783 -2.93102591
#>  [7,] -0.3381106 0.5257568  2.0128775 -0.02881796
#>  [8,]  0.8237244 0.2914427  1.6046529 -0.22979680
#>  [9,]  0.8126927 1.3577197 -0.3536178 -0.45537974