Provide random initialization for pibble model

Randomly initializes based on ALR transform of counts plus random pseudocounts uniformily distributed between 0 and 1.

random_pibble_init(Y)

Arguments

Y
matrix (D x N) of counts

Value

(D-1) x N matrix

Details

Notation: N is number of samples and D is number of multinomial categories

Examples

Y <- matrix(sample(1:100, 100), 10, 10)
random_pibble_init(Y)

#>             [,1]       [,2]       [,3]       [,4]       [,5]        [,6]
#>  [1,] -0.3301388  1.2637274 -0.1528884  0.6531966 -1.1863823 -2.24819258
#>  [2,] -0.9520309  1.3743205 -1.9395433  1.2649020 -0.8110813  0.27711268
#>  [3,] -1.5843328  1.0180952 -1.0701414  0.9902248  0.0360073 -0.02674815
#>  [4,] -0.4403705  0.9210183 -0.5247373  0.6114326 -2.0642148 -0.58834323
#>  [5,] -0.4014171  1.3282157  0.1650161 -1.2870397 -0.9614123 -0.96874569
#>  [6,] -0.6587288  1.1714799 -0.1080758  1.1857583 -0.2637393 -1.01742196
#>  [7,] -3.6297537  0.5725964 -1.4095370 -1.6923363 -1.0672673 -0.40787511
#>  [8,] -0.1787738  1.0600890  0.1014273  1.3033448 -0.1103540 -0.54636945
#>  [9,] -1.8805005 -0.4124154 -3.3799094  0.3803609 -0.4913476 -0.39716466
#>             [,7]       [,8]        [,9]       [,10]
#>  [1,]  0.3316893  0.5792171 -2.90522256 -0.09410103
#>  [2,]  0.6384333  1.5340624 -0.05128138 -0.24797553
#>  [3,]  0.6492500 -0.4476575 -1.43416910 -0.63103977
#>  [4,]  0.7618551  1.0123159 -1.80886181 -0.29850913
#>  [5,]  0.8817291  1.1201732 -1.21892192 -0.95340068
#>  [6,]  0.6196574 -1.6461313 -0.08404239 -2.00587900
#>  [7,]  0.7344748  1.6982600 -0.04243484 -1.91288365
#>  [8,]  0.2063577  1.1281820 -0.30815486 -1.21033415
#>  [9,] -0.6982748  1.5635580 -0.13096377 -0.83143706