Provide random initialization for pibble model

Randomly initializes based on ALR transform of counts plus random pseudocounts uniformily distributed between 0 and 1.

random_pibble_init(Y)

Arguments

Y
matrix (D x N) of counts

Value

(D-1) x N matrix

Details

Notation: N is number of samples and D is number of multinomial categories

Examples

Y <- matrix(sample(1:100, 100), 10, 10)
random_pibble_init(Y)

#>              [,1]       [,2]       [,3]        [,4]        [,5]        [,6]
#>  [1,] -0.24231935 -2.0451329 -1.8213889 -2.00968851 -0.20776321 -0.81574922
#>  [2,] -0.43797892 -0.1077768 -1.1272594  0.08226038 -2.91394753  0.16354158
#>  [3,] -1.42023460  0.3843112 -2.1287271 -0.37142524  0.14644789 -0.10603079
#>  [4,] -0.05155833  1.0056633 -0.1502702 -0.33784553 -0.62700311 -1.61855023
#>  [5,]  0.26573821  1.1581450  0.2952061 -0.78320252  0.02011466 -0.96935261
#>  [6,]  0.22236651 -1.0289364 -0.2606817 -0.08547448  0.08512388  0.03838981
#>  [7,] -3.16092877  0.2277096 -2.4680390  0.12356439 -1.02755693 -0.37979454
#>  [8,] -0.74348709  0.1547087  0.2843538 -0.28096773 -1.33586974 -0.58842984
#>  [9,] -0.30495504  0.9080730 -0.5451461 -1.46195554 -2.33730802 -0.30579097
#>              [,7]        [,8]       [,9]      [,10]
#>  [1,] -0.13781882 -2.64683050 -0.2659892 -4.0431102
#>  [2,] -0.83356676 -0.58119959 -1.1707708 -0.4470475
#>  [3,]  0.26507286  0.02331203 -0.8118606 -0.2855162
#>  [4,] -0.30727103 -1.10991161 -0.1692626 -1.8847652
#>  [5,] -0.34687754 -0.41307589 -0.1514538 -1.4511327
#>  [6,]  0.08648078  0.14826052 -1.8043664 -0.7441311
#>  [7,]  0.37008195  0.12848290 -0.4552256 -1.5402702
#>  [8,] -1.60979215  0.16917044 -1.6400883 -1.6211847
#>  [9,] -0.64504373 -0.52795839 -0.3602502 -0.6857193