Provide random initialization for pibble model

Randomly initializes based on ALR transform of counts plus random pseudocounts uniformily distributed between 0 and 1.

random_pibble_init(Y)

Arguments

Y
matrix (D x N) of counts

Value

(D-1) x N matrix

Details

Notation: N is number of samples and D is number of multinomial categories

Examples

Y <- matrix(sample(1:100, 100), 10, 10)
random_pibble_init(Y)

#>             [,1]       [,2]       [,3]        [,4]        [,5]       [,6]
#>  [1,]  0.2564795  0.4020805 -0.7263420  0.35888599 -1.83141559 -1.4781918
#>  [2,]  0.1308749 -0.5910188 -2.5429390  0.09552720 -0.03166671 -0.7306376
#>  [3,] -1.4967340 -0.1596885 -0.8951563 -1.24989293  0.95177772 -0.4206165
#>  [4,] -1.2432162  0.1949678 -2.9578447 -1.77045438 -0.35202180 -0.9247950
#>  [5,] -3.6991877  0.3010896 -0.5365658  0.08785387 -0.88501064  0.2424330
#>  [6,]  0.2351171 -0.8361655 -2.1002296 -0.31882805  0.66108164  0.2627952
#>  [7,] -0.4201395  0.4588161 -1.2490576  0.24324713  0.42298078  0.2978652
#>  [8,]  0.6604881 -0.4429850 -1.4594944  0.32040572  0.64412893 -0.5090139
#>  [9,] -0.5692143  0.3218052 -0.4788465 -0.18219510 -0.66629824 -0.1659370
#>             [,7]        [,8]        [,9]      [,10]
#>  [1,] -1.0655764  0.29871031 -0.25558951  0.2780007
#>  [2,] -0.6146730  0.11231564 -3.43849940  0.6276261
#>  [3,] -0.2613160 -0.01800373  0.15920309  0.7382263
#>  [4,] -3.0107526 -1.18187053 -2.99578035  1.1323288
#>  [5,] -0.8797707 -0.35876499  0.09902964 -1.3071641
#>  [6,] -1.8993427  0.13710311 -1.46800116  1.0201013
#>  [7,] -1.1369392  0.29158771  0.04334499  1.0012121
#>  [8,] -1.4463443 -0.70383301 -0.61776865  0.7535658
#>  [9,] -0.3489569 -1.12747124 -0.43375908 -0.8005397