Provide random initialization for pibble model

Randomly initializes based on ALR transform of counts plus random pseudocounts uniformily distributed between 0 and 1.

random_pibble_init(Y)

Arguments

Y
matrix (D x N) of counts

Value

(D-1) x N matrix

Details

Notation: N is number of samples and D is number of multinomial categories

Examples

Y <- matrix(sample(1:100, 100), 10, 10)
random_pibble_init(Y)

#>              [,1]        [,2]        [,3]       [,4]        [,5]       [,6]
#>  [1,] -2.09826445  0.88531473 -1.30372494 -0.5882382  0.05973214 -0.2111874
#>  [2,] -0.18470733  0.61990934 -1.27249868 -0.2515906 -0.21061782 -1.2355682
#>  [3,] -0.33894858  0.04472266 -0.34768575 -0.5735067  0.37881564 -1.9589151
#>  [4,]  0.26298594  0.20327414  0.06091647 -1.3600359  0.14786803 -0.5654040
#>  [5,]  0.08522398 -0.27047533 -0.13648772  0.2369855  0.08537134 -2.8594807
#>  [6,]  0.15173136 -0.77365757 -1.06399962  0.1572791  0.01055715 -1.0200058
#>  [7,] -0.71578613  0.78829673 -0.24485924  0.2704599 -0.52811855 -1.2668447
#>  [8,] -1.13832310 -0.73975983 -0.12695642 -0.3521450  0.44275704 -0.2206490
#>  [9,] -0.05236464 -2.43771262 -0.61716516 -3.4077168  0.48015052 -0.1354713
#>              [,7]     [,8]       [,9]       [,10]
#>  [1,]  0.01285488 1.434970  0.3453675 -0.15203596
#>  [2,]  0.48122670 1.845451  0.6199732  0.43043752
#>  [3,] -0.88511297 1.735517 -1.2176522 -1.54491933
#>  [4,] -0.31938667 2.276046 -0.7405710 -0.73942815
#>  [5,] -1.34070294 2.585650  1.0755786 -2.39634251
#>  [6,]  0.54538578 1.177561 -0.5104871 -1.23666783
#>  [7,] -3.75261980 1.788733  1.5328050  0.31953552
#>  [8,]  0.53167255 2.301594 -0.1385105  0.09713733
#>  [9,]  0.25801349 2.631811  1.0910697 -1.56718363