Provide random initialization for pibble model

Randomly initializes based on ALR transform of counts plus random pseudocounts uniformily distributed between 0 and 1.

random_pibble_init(Y)

Arguments

Y
matrix (D x N) of counts

Value

(D-1) x N matrix

Details

Notation: N is number of samples and D is number of multinomial categories

Examples

Y <- matrix(sample(1:100, 100), 10, 10)
random_pibble_init(Y)

#>              [,1]       [,2]       [,3]        [,4]        [,5]      [,6]
#>  [1,]  1.08148692 -1.8510595  0.3068992 -0.29418259 -0.51140683 2.8371918
#>  [2,]  1.00944924 -0.1756818 -0.1266309 -1.38139373 -0.88051181 1.2677315
#>  [3,] -0.26860309  0.2194223  0.6677155 -0.32189348 -2.93549212 2.3998147
#>  [4,]  0.11723758 -1.0520593 -1.3159870 -0.20385468  0.14564165 0.8408587
#>  [5,]  1.22365591  0.7755404 -1.0762734 -0.53573480 -0.20643899 0.7041391
#>  [6,]  0.95312547  0.5367709  0.7235215 -0.80394031 -0.52691156 2.8735546
#>  [7,] -0.19745168  0.1567910  0.3623839  0.15236415 -1.58116434 2.4769216
#>  [8,] -0.07002981  0.8025390  0.8516511 -3.17253229  0.03099507 2.5786081
#>  [9,] -0.10503160  0.1380126 -1.1669521  0.01799195 -1.32444186 2.1994325
#>             [,7]       [,8]       [,9]      [,10]
#>  [1,]  1.1274454  0.3146848 -0.6946108  0.3701905
#>  [2,]  1.4676757  0.5069014 -0.7311647 -3.1043483
#>  [3,]  2.2025627  0.3851819 -2.0273326 -3.3905880
#>  [4,]  1.1150827  0.3973763  0.2760187  0.7650290
#>  [5,]  2.0227500 -0.1561838  0.3284022  0.2670050
#>  [6,]  1.7082221  0.3008424 -0.5067853 -0.6885151
#>  [7,]  1.9736095 -0.3221119 -0.8895008  0.3169154
#>  [8,]  1.3160673  0.1074681  0.3994328  0.7207511
#>  [9,] -0.2679709 -1.1144212  0.2375592 -0.1586304